### 2019-05-05 #1430 @fauvernierma
Hello everyone,

As they are both experts in R and survival analysis, I think that @becarioprecario and @adibender would make appropriate reviewers for the survPen package.

Do you agree ?
