### 2019-05-05 #1430 @csoneson
@fauvernierma - thanks for these suggestions. I agree that they would fit very well as reviewers. @becarioprecario, @adibender, would you be willing to review this JOSS submission?
