### 2019-05-05 #1430 @csoneson
@fauvernierma - I had a first quick look at your repository and the paper, and I have a couple of small comments that would be good if you could fix already before we start the review:
- Please add a plain text LICENSE file (you can add it to `.Rbuildignore` to ignore it when building the package
- Please check the references in the paper. Some of them look a bit weird (e.g. the Uhry _et al_ paper have authors "Group, G.E.W" and "the, C.W.S.G.", and not all words are properly capitalized). Also make sure that a DOI is indicated for all references that have one.
