### 2019-05-06 #1430 @csoneson
@corybrunson, @seabbs - would you be interested in reviewing this submission for JOSS?
