+++ paper/paper.bib
@@ -16,7 +16,7 @@
 }
 
 @article{Esteve1990,
-	author = {Esteve, J. and Benhamou, E. and Croasdale, M. and Raymond, L.},
+	author = {Estève, J. and Benhamou, E. and Croasdale, M. and Raymond, L.},
 	title = {Relative survival and the estimation of net survival: elements for further discussion},
 	journal = {Stat Med},
 	volume = {9},
@@ -24,6 +24,7 @@
 	pages = {529-38},
 	ISSN = {0277-6715 (Print)
 	0277-6715 (Linking)},
+	DOI = {10.1002/sim.4780090506},
 	url = {https://www.ncbi.nlm.nih.gov/pubmed/2349404},
 	year = {1990},
 	type = {Journal Article}
@@ -45,6 +46,7 @@
 	number = {2},
 	pages = {363-379},
 	ISSN = {0196-5204},
+	DOI = {10.1137/0909024},
 	year = {1988},
 	type = {Journal Article}
 }
@@ -57,6 +59,7 @@
 	number = {1},
 	pages = {113-120},
 	ISSN = {0006-341X},
+	DOI = {10.1111/j.1541-0420.2011.01640.x},
 	year = {2012},
 	type = {Journal Article}
 }
@@ -73,27 +76,29 @@
 
 
 @article{Remontet2007,
-	author = {Remontet, Laurent and Bossard, Nadine and Belot, Aurélien and Esteve, Jaume and FRANCIM, French Network of Cancer Registries},
+	author = {Remontet, Laurent and Bossard, Nadine and Belot, Aurélien and Estève, Jacques and the French Network of Cancer Registries},
 	title = {An overall strategy based on regression models to estimate relative survival and model the effects of prognostic factors in cancer survival studies},
 	journal = {Statistics in medicine},
 	volume = {26},
 	number = {10},
 	pages = {2214-2228},
 	ISSN = {0277-6715},
+	DOI = {10.1002/sim.2656},
 	year = {2007},
 	type = {Journal Article}
 }
 
 
-@article{Ruppert2003,
+@book{Ruppert2003,
 	author = {Ruppert, David and Wand, Matt P and Carroll, Raymond J},
-	title = {Semiparametric Regression. Cambridge Series in Statistical and Probabilistic Mathematics 12},
-	journal = {Cambridge: Cambridge Univ. Press. Mathematical Reviews (MathSciNet): MR1998720},
+	title = {Semiparametric Regression},
+	publisher = {Cambridge University Press},
 	year = {2003},
-	type = {Journal Article}
+	DOI = {10.1017/CBO9780511755453},
+	type = {Book}
 }
 @article{Uhry2017,
-	author = {Uhry, Z. and Bossard, N. and Remontet, L. and Iwaz, J. and Roche, L. and Group, Grell Eurocare- Working and the, Censur Working Survival Group},
+	author = {Uhry, Z. and Bossard, N. and Remontet, L. and Iwaz, J. and Roche, L. and the Grell Eurocare-5 Working Group and the Censur Working Survival Group},
 	title = {New insights into survival trend analyses in cancer population-based studies: the SUDCAN methodology},
 	journal = {Eur J Cancer Prev},
 	volume = {26 Trends in cancer net survival in six European Latin Countries: the SUDCAN study},
@@ -109,7 +114,7 @@
 	author = {Wood, Simon N},
 	title = {Generalized additive models: an introduction with R},
 	publisher = {Second Edition, Chapman and Hall/CRC},
-	ISBN = {1498728340},
+	ISBN = {9781498728331},
 	year = {2017},
 	type = {Book}
 }
@@ -123,6 +128,7 @@
[…]
