### 2019-05-06 #1430 @fauvernierma
I have added a LICENSE file and all references should be in order now

Thank you
