### 2019-05-06 #1430 @corybrunson
@csoneson i would be glad to review. I would expect to finish after one week. (I have only introductory training in survival analysis but the topic interests me.)
