### 2019-05-07 #1430 @seabbs
@csonesn I would also be happy to review.

Not an expert in survival analysis but I have used `mgcv` fairly regularly (according to vignettes it has a comparable framework).
