### 2019-05-07 #1434 @whedon
[REVIEW]: survPen: an R package for hazard and excess hazard modelling with multidimensional penalized splines
**Submitting author:** @fauvernierma (<a href="http://orcid.org/0000-0002-3809-8248">Mathieu Fauvernier</a>)
**Repository:** <a href="https://github.com/fauvernierma/survPen" target ="_blank">https://github.com/fauvernierma/survPen</a>
**Version:** v1.2.0
