### 2019-05-14 fauvernierma/survPen#2 @seabbs
License needs package name etc.
As part of this review openjournals/joss-reviews#1434

Your license file needs to refer specifically to your package. See [here]( https://github.com/seabbs/getTBinR/blob/c43f2dc20ac036d6cc93ec32bb9b33f9b4399f28/LICENSE#L634) for an example.
