### 2019-05-14 fauvernierma/survPen#3 @seabbs
Package release versioning
As part of this review openjournals/joss-reviews#1434

It would be great to see explicit versioning of your package releases. This can be done with a tool like [Zenodo](https://zenodo.org) and Github tags. This is also nice to have as it gives you a software DOI to refer to on top of the JOSS paper DOI (allowing the actual version of the package used to be referenced)

As for the license see [here](https://github.com/seabbs/getTBinR) for an example of this (apologies for the self-referencing).

For Zenodo the steps are as follows:

* Sign up to Zenodo and turn of archiving for this repository
* Make a release using GitHub tags - this should then be automatically tagged by Zenodo
* Add the markdown badge to your readme (either using the version badge or the most up to date DOI badge)
* Continue to make tags as new versions of the package are released (perhaps adding your news into the details section).
