### 2019-05-14 fauvernierma/survPen#4 @seabbs
Authorship
As part of this review openjournals/joss-reviews#1434

I am only seeing GitHub commits from @fauvernierma?

From the paper I am assuming that the other authors are thesis advisors etc. - could you please clarify their individual contributions.

In a normal paper, this is the sort of thing that would be listed at the end but I am not sure that JOSS stipulates this.
