### 2019-05-14 fauvernierma/survPen#5 @seabbs
Paper: Statement of need + who is this for
As part of this review openjournals/joss-reviews#1434

At the moment I feel like readers need to look over the whole paper to get a good idea of why the package is needed. It would be great if the need could be a little clearer in the first paragraph. There may also be an issue for people who don't know why splines are a good choice.

So my reading of the need for this package is that,

1. In survival analysis, we need to fit models that account for various confounders with potential complex non-linear effects that are not known a priori
2. Penalised regression splines represent an appealing solution (explain in more detail?)
3. Current implementations are slow and/or unstable. `survPen` represents a fully automated, fast, and stable tool set.

Could you condense this (improve my blatherings!) for the opening paragraph of the paper?

I am also not seeing anything referencing who this is for explicitly - perhaps add a sentence. The examples etc. make it clear that it can be used by epidemiologists with survival data but potentially has quite a few other uses?
