### 2019-05-14 fauvernierma/survPen#6 @seabbs
References
As part of this review openjournals/joss-reviews#1434

* Your second reference has no DOI
* [*optional*] For the links within the paper consider using markdown links rather than explicitly writing out (cleaner)
