### 2019-05-14 fauvernierma/survPen#7 @seabbs
Community guidelines
As part of this review openjournals/joss-reviews#1434

As an open source tool this needs clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

A good way to do this is to steal a page from the tidyverse toolbox.

I would like to see the following,

1. Information on how to contribute in the README.
1. Link to code of conduct in the readme
1. Code of conduct, contributing and support either in the readme or as separate markdown documents.

See [here](https://github.com/seabbs/getTBinR/tree/master/.github) and [here](https://github.com/seabbs/getTBinR/blob/master/README.md) (again! :smile:) for my tidyverse inspired approach.
