### 2019-05-14 fauvernierma/survPen#8 @seabbs
Automated tests
As part of this review openjournals/joss-reviews#1434

Looking at the testing script in the readme it looks like the code used in the vignette (I may be wrong about that!). I ran it through but as the expected output is not listed it is difficult to know if everything is working as expected!

Automated tests can be a great framework for

1. Writing tests that others can use
2. Catching errors as they slip in
3. Establishing if all of the software is being tested.

I would really like to see a suite of [`testthat`](https://testthat.r-lib.org) tests (adapted from the current testing script), ideally, run on each commit using something like [travis](https://travis-ci.org). A test coverage report from [codecov](https://codecov.io) would also be great.

To get all this working you need to do the following,

1. Look at the `testthat` documentation and write unit tests for each function (based on your current testing script). See [here](https://github.com/seabbs/getTBinR/tree/master/tests/testthat) (😸) for some examples. These can then be run from the package root using `devtools::test()`.
2. Set up travis (optionally also appveyor for windows) to run these tests - using [`usethis`](https://usethis.r-lib.org/reference/ci.html). See [here](https://github.com/seabbs/getTBinR/blob/master/.travis.yml) for an example.
3. Set up codecov to generate reports - [`usethis`](https://usethis.r-lib.org/reference/use_coverage.html?q=use%20_%20covera). See [here](https://github.com/seabbs/getTBinR/blob/master/.travis.yml) for an example.
4. Add markdown badges with test status and coverage to your README.  See [here](https://github.com/seabbs/getTBinR/blob/master/README.Rmd) for an example.
5. Get all tests passing and code coverage to >80%.
