### 2019-05-14 fauvernierma/survPen#9 @seabbs
Documentation
As part of this review openjournals/joss-reviews#1434

As I said in the review thread the vignette for this package is fantastic. Your within R function documentation is also very good (👍) - `survPen` has some of the best documentation I have ever seen. However, I think a few small changes could lead to some big usability improvements.

1. Flesh out the readme. I really like to see a detailed readme with some quick details about the package.
   * Why this package
   * Installation (you have this :smile:)
   * Some quick start information. The key functions and/or an example of using the package to solve some issue. This is effectively selling the package to the user.
   * As mentioned in other issues add test badges along with other badges giving release information etc. (*optional*). See [here](https://github.com/seabbs/getTBinR/blob/master/README.Rmd).
2. (*optional*) Personally I think I would split the current readme into separate documents - as it currently stands it is a little intimidating!
3. I would really like to see an example of using this package versus using some of the other available packages to solve the same issues. This would really help the potential user to understand why what your offering is so much better. It would also act as a nice check that this packages produces similiar results (hopefully much faster) to other more established packages.
4. Package description needs to be more detailed. Consider using the first paragraph summarising the package from the readme.
5. Package website. I find looking at package documentation + getting a feel for how the package works to be much easier if the package has a website (i.e [`getTBinR`](https://www.samabbott.co.uk/getTBinR/)). This is particularly helpful for understanding what functions the package offers (similar to the manual PDF). Setting up a website is very straightforward so I would very much recommend it. Steps,
   * Install `pkgdown` and see [docs](https://pkgdown.r-lib.org).
   * Set up the  `_pkgdown.yml` either using [`usethis`](https://usethis.r-lib.org/reference/use_pkgdown.html) or with an [example](https://github.com/seabbs/getTBinR/blob/master/_pkgdown.yml).
   * Build the website using `pkgdown::build_site()` (can also build and deploy on travis - see [here](https://pkgdown.r-lib.org/articles/pkgdown.html#configuration).
   * Iterate and improve.

I realise this looks like a lot to do but really only small changes are needed. In general docs are great!
