### 2019-05-14 fauvernierma/survPen#10 @seabbs
Performance claims
As part of this review openjournals/joss-reviews#1434

As mentioned in #9 I would like to see a comparison of `survPen` to other software. How much faster/ more stable is it. In what circumstances is it better (or is it ever worse)? When does it agree/disagree

Some of these may be in the methods paper that you are submitting but without a preprint (is there one) I can't assess this.

A good target would be to expand on this "The package mgcv can then be used to fit penalized hazard models (Remontet et al. 2018). The problem with this option is that the setup is rather complex and the method cannot be used on very large datasets for computational reasons." from the vignette with an example.

It would also be great to see `survPen` compared to other non-spline based methods (and get to see the big improvements it gives!).
