### 2019-05-14 #1434 @seabbs
Hi @fauvernierma,

First off this package looks great and I particularly like the detailed vignette. 👍

Broken down comments by section.

- [ ]  License needs references to your actual package in the appropriate places. fauvernierma/survPen#2
- [ ] Whilst your released version (i.e CRAN) matches your dev version (GitHub) I would ideally see some kind of explicit archiving of given releases. See fauvernierma/survPen#3 for details - this is not an acceptance blocker.
- [ ] Authorship - I am only seeing commits from the first author.  Is a contribution statement in the paper a good idea @csoneson. See fauvernierma/survPen#4

- [ ] Statement of need - needs some fine tuning. See  fauvernierma/survPen#5
- [ ] References - missing DOI. See fauvernierma/survPen#6

- [ ] Statement of need - as for fauvernierma/survPen#5.  This needs to be in the package description and the readme. Vignette is very clear but lacks a quick intro as to why you want to use this package.
- [ ] Documentation - see  fauvernierma/survPen#9
- [ ] Tests - need automated unit tests.  See fauvernierma/survPen#8
- [ ] Community guidelines - missing. See fauvernierma/survPen#7

- [ ] Performance - paper says `survPen` is faster + more stable than alternatives. An example of this is needed. See fauvernierma/survPen#10
- [ ] Functionality - on the surface this all looks great. Unfortunately, I am not qualified to assess the underlying methods. As above I would like to see a comparison to alternative software solutions (even solutions not using splines) to see the improvement this method gives.

## Misc

* The code base looks robust and well implemented.
* `survPen` has minimal dependencies which makes installation etc. a breeze.
- [ ] [*optional*] You state in the description that at least R `3.5.0` is required. Is this true? It might be good to allow for some earlier R versions if possible
- [ ] [*optional*] I prefer each function (or group of functions) to have their own R script.

Overall `survPen` looks like a great package and a useful contribution to the R universe 👍. I am looking forward to conducting future analyses using it. However, there are several outstanding issues with testing, documentation and performance/functionality (outlined above) that need resolving before I can recommend this package. Happy to discuss these points further in the relevant issues and provide additional support as required :smile:
