### 2019-05-14 fauvernierma/survPen#11 @seabbs
Citing
As part of this review openjournals/joss-reviews#1434

I recommend adding citation information to the package (will need updating upon JOSS acceptance) and to the readme.

Examples for [package citation](https://github.com/seabbs/getTBinR/blob/master/inst/CITATION) and [readme](https://github.com/seabbs/getTBinR/blob/c43f2dc20ac036d6cc93ec32bb9b33f9b4399f28/README.Rmd#L126)

You may also want to say when the package should be referenced versus the methods paper currently under review (or both). If both you need to more explicitly point to the methods paper as something that needs to be cited.
