### 2019-05-14 fauvernierma/survPen#12 @seabbs
Binder
As part of this review openjournals/joss-reviews#1434

I would love to see a launch with binder option to allow people to start playing around with the package as quickly as possible.

"Binder allows you to create custom computing environments that can be shared and used by many remote users. It is powered by BinderHub, which is an open-source tool that deploys the Binder service in the cloud. One-such deployment lives here, at mybinder.org, and is free to use."

See [docs](https://mybinder.readthedocs.io/en/latest/) for more information. Set up can be a bit of a fiddle if not used to docker so you may want to borrow from [another](https://github.com/seabbs/getTBinR/blob/master/binder/Dockerfile) implementation.

Once everything is set-up just added a button to the readme. Along the lines of,
