### 2019-05-14 c535fdb
Update LICENSE, issue #2
response to issue: License needs package name etc. #2
