### 2019-05-14 #1434 @fauvernierma
Thank you very much for @seabbs for the review

@csoneson, should I go to "pull request" in github to address each issue or can I just edit the file ?
And also am I authorized to close the issues myself when I consider that they have been addressed ?

Thanks
