### 2019-05-14 fauvernierma/survPen#3 @fauvernierma
Thank you @seabbs for clearly explaining the different steps.

As I understand from the other JOSS reviews that I read, the Zenodo archive is usually done on acceptance. So should I do it now ?
