### 2019-05-14 fauvernierma/survPen#3 @seabbs
No problem.

Hmm I am not sure that is the case but you could be correct.

It is generally good practise to have set-up. When the paper is accepted you can then tag and archive a new version. I think that is the best way to go.

Basically, you should be tagging every major release. People can then reference the version of your software that they are using.
