### 2019-05-14 #1434 @csoneson
Thanks for your review @seabbs! Regarding the author contribution statement, we currently don't require that. Ultimately, authors are responsible for deciding which contributions are sufficient for co-authorship (see more at https://joss.readthedocs.io/en/latest/review_criteria.html#authorship). @fauvernierma, since it's difficult to judge from the commit history in your repository - could you please confirm that all the authors qualify for inclusion with respect to these criteria, to address @seabbs's comment?
