### 2019-05-14 #1434 @csoneson
@fauvernierma: You don't have to make a separate pull request for each fix. It may be useful to keep each fix in a separate commit, and refer to the commit in the corresponding issue. This would keep things modular and make it easy for the reviewers to see what you changed. You can also close the issues in your repository - the evaluation from the reviewers will be recorded in the checklist in the beginning of this issue; this is where they will record that a fix is satisfactory to address their comments.
