### 2019-05-14 #1434 @fauvernierma
Ok, thanks @csoneson .

@seabbs, as for the authorship, the four other authors/contributors and I are part of a survival working group within our research department. Two of them (Laurent Remontet and Nadine Bossard) are indeed my thesis advisors.
Laurent Roche and Laurent Remontet are cited as authors as they intervened in the core development of the package (especially in linear algebra and matrix derivatives).
As for Zoé Uhry and Nadine Bossard, they contributed to the writing of the documentation and gave important pieces of advice relative to the package functionalities.

I hope this addresses your comment @seabbs
