### 2019-05-14 fauvernierma/survPen#10 @fauvernierma
There is indeed an extensive comparison to other existing packages in the submitted paper so I have to see which elements I am allowed to make available online (for copyright reasons)
