### 2019-05-14 fauvernierma/survPen#6 @fauvernierma
If I am right, the missing DOIs only concern Fauvernier2019, R and Wood2017 and I think it is normal that they do not have one : the first one is my submitted article, the second one is for the R software and the third has a ISBN number

Is there something that I am missing ?

Concerning the links, isn't the @ considered a markdown link ?

Thank you and sorry If my questions seem a bit naive
