### 2019-05-14 fauvernierma/survPen#6 @seabbs
I only meant Fauvernier2019 for the DOI but based on what you have said I guess this is under review. It seems a shame that this isn't available to support the package as it seems so critical. As I mentioned in the other issue have you considered preprinting.

For the links I meant the GitHub links etc not the referencing - sorry for the confusion. I was suggesting that you use markdown [links](https://github.com) rather than pasting the full link (i.e https://github.com). This is more of a style consideration so don't worry if you disagree.

If not open to preprinting then I am happy for you to close this issue.
