### 2019-05-14 1e206d5
Changing format of vignette citations
