### 2019-05-14 fauvernierma/survPen#6 @fauvernierma
Ok, thank you

I admit that the co-authors and I have not considered preprinting but hopefully the article will be accepted soon and I will be able to properly cite it.

Concerning the links, I made the following commit 1e206d5383f759e3cfe1f6417669a4e3bf9f6abc
