### 2019-05-14 fauvernierma/survPen#6 @seabbs
Okay great - happy to close this issue.

Just a final note: I would really strongly urge you to preprint if at all possible. You can use [this](http://sherpa.ac.uk/romeo/index.php) to check the preprint policies at your target journal.
