+++ compar/Lexis.R
@@ -1,0 +1,232 @@
+#---------------------------------------------------------------------
+# lexis
+
+#----------------------------------------------------------------------
+# Corresponds to W.Lexis
+# Original by David Clayton, approx. 1999.
+#
+# Modified version by BxC, July 2002:
+#   Returns numerical left endpoints equal to breaks instead of factors.
+#   Small random quantity added to origins to break ties between cuts.
+#----------------------------------------------------------------------
+
+lexis <- 
+function( entry = 0,
+           exit, 
+           fail,
+         origin = 0,
+          scale = 1,
+         breaks,
+        include = NULL,
+           data = NULL ) 
+{
+
+# If there is a dataframe argument, attach it so that arguments are matched
+  if (!is.null(data)) {
+    attach(data, 2)
+    on.exit(detach(pos=2))
+  }
+  a.en <- attributes(entry)
+  a.ex <- attributes(exit)
+  a.fa <- attributes(fail)        
+  a.en$names <- NULL
+  a.ex$names <- NULL
+  a.fa$names <- NULL
+
+  n <- length(exit)
+  if (length(entry) == 1) 
+    entry <- rep(entry, n)
+  if (length(entry)!=n) 
+    stop("wrong argument length (entry)")
+  if (length(fail)!=n)
+    stop("wrong argument length (fail)") 
+  if (any(exit<entry))
+    stop("exit time < entry time")
+
+# Is there more than one time scale?
+  if(!is.list(origin)) {
+    if (length(origin)!=1 && length(origin) != n)
+      stop("illegal argument (entry)")
+    if(!is.numeric(breaks))
+      stop("illegal argument (breaks)")
+    if(!is.numeric(scale) || length(scale)!=1)
+      stop("illegal argument (scale)")
+    origin <- list(Time = origin)
+    breaks <- list(Time = breaks)
+    names(scale) <- "Time"
+    }
+  if (!is.list(breaks))
+    stop("breaks argument should be a list")
+
+# Find time scale names in breaks argument and check, otherwise assign
+  tnames <- names(origin)
+  nt <- length(tnames)
+  bnames <- names(breaks)
+  if (is.null(bnames)) {
+    if (length(breaks) != nt) 
+      stop("wrong argument length (breaks)")
+    names(breaks) <- tnames
+  }
+  else { 
+    if(any(sort(bnames) != sort(tnames)))
+      stop("breaks and origin names don't match")
+  }
+
+# Find names of scale argument
+  bnames <- names(scale)
+  if (is.null(bnames)) {
+    if (length(scale) == 1 && nt != 1)
+      scale <- rep(scale, nt)
+    names(scale) <- tnames
+  }
+
+# Add a small random quantity to the origins to break timescale ties
+  for (s in tnames) {
+    origin[[s]] <- origin[[s]] + runif(1)*scale[[s]]/10^6
+  }
+
+# Covariates
+  if (!is.null(include)) {
+    covexp <- substitute(include)
+    if (mode(covexp)=="name") {
+      include <- list(include)
+      names(include) <- as.character(covexp)
+    }
+    else if (mode(covexp)=="call" && covexp[[1]]=="list") {
+      nc <- length(covexp)
+      cnames <- names(covexp)
+      if (!is.null(cnames)) {
+        for (i in 2:nc) {
+          if(cnames[i] == "") {
[…]
