### 2019-05-15 fauvernierma/survPen#10 @fauvernierma
Hi @seabbs,

I just added a [compar](https://github.com/fauvernierma/survPen/tree/master/compar) folder
that illustrates a comparison between survPen, mgcv and rstpm2 (another reference package in the area)

I hope this will answer your question about perfomance claim (tell me if that is not the case)

Also, I am not sure that adding this folder is the best way to go, maybe I should add this to the vignette ?

Thank you :)
