### 2019-05-15 fauvernierma/survPen#11 @fauvernierma
I added a package citation 72852b9ec56fec8180b8c4c8f6c7dc985cd5b10f (but I am not sure of how to deal with group author) and a citation header in the readme 04183aed8cb56a8a17b89f0bf9e5841e4dd26523
