### 2019-05-15 0b55d35
Create CONTRIBUTING.md
