+++ .github/CONTRIBUTING.md
@@ -1,0 +1,34 @@
+# Contributing to survPen
+
+This outlines how to propose a change to survPen.
+
+### Fixing typos
+
+Small typos or grammatical errors in documentation may be edited directly using
+the GitHub web interface, so long as the changes are made in the _source_ file.
+
+*  YES: you edit a roxygen comment in a `.R` file below `R/`.
+*  NO: you edit an `.Rd` file below `man/`.
+
+### Prerequisites
+
+Before you make a substantial pull request, you should always file an issue and
+make sure someone from the team agrees that it’s a problem. If you’ve found a
+bug, create an associated issue and illustrate the bug.
+
+### Pull request process
+
+*  We recommend that you create a Git branch for each pull request (PR).  
+*  We use [roxygen2](https://cran.r-project.org/package=roxygen2), with
+[Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/markdown.html), 
+for documentation.  
+*  For user-facing changes, add a bullet to the top of `NEWS` below the
+current development version header describing the changes made followed by your
+GitHub username, and links to relevant issue(s)/PR(s).
+
+### Code of Conduct
+
+Please note that the survPen project is released with a
+[Contributor Code of Conduct](https://github.com/fauvernierma/survPen/blob/master/.github/CODE_OF_CONDUCT.md). By contributing to this
+project you agree to abide by its terms.
+
