+++ .github/SUPPORT.md
@@ -1,0 +1,22 @@
+# Getting help with survPen
+
+Thanks for using survPen. Before filing an issue, there are a few places
+to explore and pieces to put together to make the process as smooth as possible.
+
+Where should you ask for help ? 
+
+  * If it's a question: start with [community.rstudio.com](https://community.rstudio.com/), 
+    and/or StackOverflow. There are more people there to answer questions.  
+  * If it's a bug: you're in the right place, file an issue.  
+  * If you're not sure: let the community help you figure it out! If your 
+    problem _is_ a bug or a feature request, you can easily return here and 
+    report it. 
+
+Before opening a new issue, be sure to [search issues and pull requests](https://github.com/fauvernierma/survPen/issues) to make sure the 
+bug hasn't been reported and/or already fixed in the development version. By 
+default, the search will be pre-populated with `is:issue is:open`. You can 
+[edit the qualifiers](https://help.github.com/articles/searching-issues-and-pull-requests/) 
+(e.g. `is:pr`, `is:closed`) as needed. For example, you'd simply
+remove `is:open` to search _all_ issues in the repo, open or closed.
+
+Thanks for your help!
