### 2019-05-15 8a72018
Update README.md
Adding Contributing section
