### 2019-05-15 fauvernierma/survPen#7 @fauvernierma
I added a Contributing section 8a72018ea3852dae65949c0a6195db4056729282 in the readme as well as the required [files](https://github.com/fauvernierma/survPen/tree/master/.github)
