### 2019-05-15 fauvernierma/survPen#11 @seabbs
That looks good to me. Do you want people who use the package to also cite the methods paper - could be a good idea to mention in the readme.

* [*optional*] If you change your README to a Rmd rather than a MD you can print the output of the citation. Target output needs to be a github_document.
