### 2019-05-15 fauvernierma/survPen#7 @seabbs
Contributing section looks good - are you interested in user submitted data? If not maybe delete that part!

Files all look good 👍

By the way, it might be a good idea to increment your package version for all these changes and add the updates to your news file under the new version title.
