### 2019-05-15 fauvernierma/survPen#10 @seabbs
Thanks for adding this - it goes a long way to addressing my concerns. 😄

I think this would be great as a vignette - new users will probably want this information. Having it in the package documentation is also a good idea as it will likely change over time as everything gets updated.

If possible a comparison with `mgcv` using some of the new approaches would be good. Can you use `bam` (and its performance features) for your models?
