### 2019-05-20 #1434 @corybrunson
@fauvernierma this might not be appropriate to raise as an issue on the code repo: Could you explain the value `GPL (>= 2)` in the `DESCRIPTION` file? The `LICENSE` file describes GPL-3, which is acceptable to CRAN (see [here](https://www.r-project.org/Licenses/)), but i understand GPL-3 to be slightly more restrictive than GPL-2 and i haven't seen this notation before.
