### 2019-05-20 fauvernierma/survPen#10 @fauvernierma
As you noted, the fact that the method article is not accepted yet is a bit embarrassing. In this article, survPen is validated by simulation and is compared to other existing approaches, demonstrating its superiority regarding statistical performances and/or computational time performance.

Therefore, the way I see it, survPen’s guarantee is that article (and hopefully people will soon be able to refer to it). Since the article is not yet available, I understand that you needed to make sure that the package actually did what it claimed to do. However, I do not feel comfortable referring to other existing approaches to defend survPen.

Indeed, mgcv and rstpm2 are the only other options to fit multidimensional penalized hazard and excess hazard models. However, rstpm2 can lead to severe instability (cf the comparison I provided) and even negative hazard estimation (so I am really not comfortable with referring to it).
As for mgcv, no matter how well it is programmed (and we both know it is state of the art in that matter), it is not designed to fit hazard models, let alone excess hazard models (the only exception being the new coxph family but this approach does not allow modelling the effect of time). Survival models with mgcv are only possible because of a useful old trick that allows using a Poisson model to fit a survival model; this require to split the data into very small interval and corresponds to a ‘point-milieu’ approximation of the cumulative hazard (numerical approximation of the integral).

However, this tricks comes at great costs:
•	Using the lexis function to split the data
•	Using a user-defined family (instead of Poisson) in order to deal with expected mortality rates when we want to fit excess hazard models
•	This user-defined family has to be changed when the fitting algorithm is changed. For example, the family we used to use with the glm function had to be modified in order to fit penalized excess hazard models using gam.

That being said, I can answer your bam question :)
Bam can be used to fit hazard models and I have to admit that its speed and stability are amazing. It comes toe to toe with survPen in the provided example and can be even faster with large datasets under some circumstances. However, the user-defined family used to fit excess hazard models with gam is no longer supported by bam. Maybe there is a way to make it work but if it exists it must be pretty complicated to find.
Besides, even if it were possible, that would not change the fact that the data has to be split before fitting. In the provided example, this is ok because the maximum follow-up time is 5 years and the split bands are: bands=c(seq(0,1,by=0.05),seq(1.1,5,by=0.1)). As you can see, the split depends on the maximum follow-up time, and the split data can then easily get enormous when the maximum time is 20 or even 50 years (this is the case when you study the time-to-death for some chronic diseases like multiple sclerosis for example). In contrast, survPen is not affected by this.
Thus, by nature the split+Poisson approach will always be more complicated to set up and more inefficient than a direct approach as the one proposed by survPen (by the way, the improvements proposed by Simon Wood with the bam function could be implemented in survPen in a future release.)

I know this is probably not the answer you were expecting and it is probably way too long but I hope you understand my concerns about having to refer to other approaches in order to promote survPen.

Thank you
