### 2019-05-20 #1434 @fauvernierma
Hello @corybrunson
As the new LICENSE file for survPen now uses GPL-3 (thanks to point raised by @seabbs), this old notation is going to change in the DESCRIPTION file.
