### 2019-05-20 fauvernierma/survPen#9 @corybrunson
I agree with most of what @seabbs has suggested, but i'd like to expand on (2): The original long vignette builds incrementally upon the analysis of a single data set, which i find useful, and i would suggest keeping something like it as a technical overview. It would be extremely helpful to _also_ include smaller vignettes that allow users to quickly learn the essentials of specific techniques.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1434)._
