### 2019-05-20 61cf8d3
Update test_survPen.r
