+++ tests/test_survPen.r
@@ -4,29 +4,49 @@
 
 
  library(survPen)
+ library(splines)
  data(datCancer) # simulated dataset with 2000 individuals diagnosed with cervical cancer
+ 
+ 
+ cat("\n","______________________________________________________________________________________________","\n")
+ cat("\n","survPen test code","\n")
+ cat("\n","______________________________________________________________________________________________","\n")
+
  
  #-------------------------------------------------------- example 0
  # Comparison between restricted cubic splines and penalized restricted cubic splines
+ cat("\n","______________________________________________________________________________________________","\n")
+ cat("\n","example 0","\n")
+ cat("\n","Comparison between restricted cubic splines and penalized restricted cubic splines","\n")
+ 
 
- library(splines)
 
  # unpenalized
  f <- ~ns(fu,knots=c(0.25, 0.5, 1, 2, 4),Boundary.knots=c(0,5))
 
  mod <- survPen(f,data=datCancer,t1=fu,event=dead)
 
+ cat("\n","--------------------------------","\n")
+ cat("\n","summary of the unpenalized model","\n")
+ print(summary(mod))
+ 
  # penalized
  f.pen <- ~ smf(fu,knots=c(0,0.25, 0.5, 1, 2, 4,5)) # careful here: the boundary knots are included
 
  mod.pen <- survPen(f.pen,data=datCancer,t1=fu,event=dead)
 
+ cat("\n","--------------------------------","\n")
+ cat("\n","summary of the penalized model","\n")
+ print(summary(mod.pen))
+ 
+ 
  # predictions
 
  new.time <- seq(0,5,length=50)
  pred <- predict(mod,data.frame(fu=new.time))
  pred.pen <- predict(mod.pen,data.frame(fu=new.time))
 
+ pdf("compar_unpen_pen.pdf",height=5,width=6)
  par(mfrow=c(1,1))
  plot(new.time,pred$haz,type="l",ylim=c(0,0.2),main="hazard vs time",
  xlab="time since diagnosis (years)",ylab="hazard",col="red")
@@ -34,13 +54,15 @@
  legend("topright",legend=c("unpenalized","penalized"),
  col=c("red","blue3"),lty=rep(1,2))
 
+ dev.off()
  
 
  #-------------------------------------------------------- example 1
  # hazard models with unpenalized formulas compared to a penalized tensor product smooth
-
- library(survPen)
- data(datCancer) # simulated dataset with 2000 individuals diagnosed with cervical cancer
+ cat("\n","______________________________________________________________________________________________","\n")
+ cat("\n","example 1","\n")
+ cat("\n","hazard models with unpenalized formulas compared to a penalized tensor product smooth","\n")
+ 
 
  # constant hazard model
  f.cst <- ~1
@@ -66,7 +88,6 @@
 
  # cubic B-spline of time with a knot at 1 year, linear effect of age and time-dependent effect
  # of age with a quadratic B-spline of time with a knot at 1 year
- library(splines)
  f.spline.inter.age <- ~bs(fu,knots=c(1),Boundary.knots=c(0,5))+age+
  age:bs(fu,knots=c(1),Boundary.knots=c(0,5),degree=2)
  # here, bs indicates an unpenalized cubic spline
@@ -92,6 +113,7 @@
  
  lwd1 <- 2
  
+ pdf("compar_several_mods.pdf",height=5,width=6)
  par(mfrow=c(1,1))
  plot(new.time,pred.cst$haz,type="l",ylim=c(0,0.2),main="hazard vs time",
[…]
