### 2019-05-20 fauvernierma/survPen#8 @fauvernierma
@seabbs, I have tried the testhat approach and although I found it interesting, when it comes to survPen, I have two main concerns with it:
•	I get the interest of testhat when the tests are created as the code is being developed but now that the package is fully operational, it feels a bit unnatural to try and develop the tests
•	In the survPen package, I basically have one function: survPen. Therefore, it feels quite clumsy to try and test any little details of each inner function

Would it not be more appropriate to a text file with the expected output and pdfs for expected graphs? This is what I did here: https://github.com/fauvernierma/survPen/tree/master/tests

Thank you
