+++ README.Rmd
@@ -1,3 +1,7 @@
+---
+output: github_document
+---
+
 # survPen
 ## hazard and excess hazard modelling with multidimensional penalized splines
 [![DOI](https://zenodo.org/badge/181266005.svg)](https://zenodo.org/badge/latestdoi/181266005)
