### 2019-05-20 fauvernierma/survPen#10 @seabbs
Thanks for all of this useful information! The poisson model trick was not something I had come across.

It is potentially problematic that the methods paper is not available as without it makes the software difficult to both review and use.  When accepted will the paper be open or closed access? If closed source then I can imagine this bring further problems (your users will potentially not be able to confirm your software actually works etc. if not academics with access)

I'm not really sure I do understand your concerns I am afraid. The way I see it the logical thing to do would be to have a section in your current vignette, or (my preferred option) in another vignette, that details what was previously available (i.e `mgcv` etc.) and then walks through an example comparing these approaches to `survPen`.  This could be very similar to your current script in `compar`.

If you think of an outsider looking at your work they need to be able to easily see:

* Does this package work?
* Is this package better/faster/comparable to other tools I may know?
* When should I use this tool over other tools.

I don't think you can answer these questions by just linking to an unavailable methods paper, especially if that paper won't be open access in the future.

Please correct me if I am misunderstanding something here! 😄
