### 2019-05-20 fauvernierma/survPen#13 @corybrunson
inline code formatting
The [thorough vignette](https://htmlpreview.github.io/?https://github.com/fauvernierma/survPen/blob/master/inst/doc/survival_analysis_with_survPen.html) contains much _italicized inline code_ that would be easier to read / scan / refer back to if formatted as `verbatim inline code`. It would also help easily distinguish the code from the mathematical expressions. This is a minor point, i know, but good formatting becomes all the more valuable in lengthier documents!

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1434)._
