### 2019-05-20 0e06ca7
added method article citation
