### 2019-05-20 fauvernierma/survPen#8 @seabbs
The updated tests look improved but I don't think they are sufficient.

These are the JOSS [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html):

**Good:** An automated test suite hooked up to an external service such as Travis-CI or similar
**OK:** Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g. a sample input file to assert behaviour)
**Bad** (not acceptable): No way for you the reviewer to objectively assess whether the software works

As a minimum, I need to see considerably more documentation in the current testing script. This needs to outline:

* What is being tested.
* What is the expected result.
* What is the actual result.
* What functionality is not tested.

The advantages of automated unit tests for both myself and your future users are as follows,

* Can assess what is being tested and what is not being tested. i.e how confident are we that everything works as stated.
* Using CI lets the current status of tests be quickly evaluated. As it currently stands all users must run your tests for themselves and compare.
* Testing each small section of your package builds confidence in the greater whole. What happens if there is a small error in one of your sub-functions that causes slight numerical discrepancies?
* Shows that software best practises are being followed meaning that this package is worth spending time to learn.

For the developer

* Adding a test when bugs are fixed prevents those bugs from reoccurring.
* Can add new features without accidentally breaking old features.
* Helps think directly about how each feature actually works and checks gives the results your expect.
* Allows areas of code that are not being tested to be identified.

I realise that proper testing seems like a lot of work for nothing but it will definitely pay itself off in the terms of user trust, and reducing any future development time. As the JOSS guidelines state adequate manual tests will be acceptable but personally I would be very wary of any package without proper tests if it did not already have a large user base (i.e `mgcv`).
