### 2019-05-20 fauvernierma/survPen#11 @seabbs
Nice.

It looks like you need to rebuild the package and knit the readme again as it currently shows the wrong citation.

There is also a problem with the formatting your md - perhaps from kniting? it is showing the ouput type at the beginning.
