### 2019-05-20 #1434 @seabbs
@csoneson - there is a methods paper under review for this package that is not available to the reviewers/users. From the references to it in the documentation (and from what @fauvernierma  has said) it contains much of the detail confirming the package functionality/performance.

What is the best approach to avoid replicating work for  @fauvernierma but still providing reviewers/users with assurances that everything works as advertised?
