### 2019-05-20 fauvernierma/survPen#14 @corybrunson
documentation examples
As noted in #9 the documentation is impressive. Regarding the exported documentation (for `help()`), the primary five functions (`survPen()` and the spline constructors) include several examples, but most of the helper functions (by which i mean, all other functions) do not. And many of these functions rely on a deep understanding of the package internals to understand and test.

This leads me to ask: Is the helper function documentation intended for developers or for users? If for developers, then should it be exported? If for users, then should examples not be included?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1434)._
