### 2019-05-20 fauvernierma/survPen#8 @corybrunson
I'd like to second @seabbs's previous comment. I'm able to manually run the test code and compare it to the provided output and PDFs, but it's a time-consuming and error-prone process.

While using a testing package and continuous integration service is optional, they make the adding and editing of tests much more efficient.

As an aside, whether to include tests for the helper functions may depend on whether they are intended to be user-facing (see #14).
