### 2019-05-20 4ccb022
adding testthat
