### 2019-05-20 11da2cb
added test files for testthat
