### 2019-05-20 60b42e0
Create .travis.yml
