### 2019-05-20 37bb9c2
survPen depends on R version >= 3.2.0 instead of 3.5.0
