### 2019-05-20 fauvernierma/survPen#15 @corybrunson
statement of need (readme)
The package documentation (`?survPen`) and JOSS paper provide clear context and motivation for the package. To help users who first encounter the package on GitHub or elsewhere (and to satisfy JOSS guidelines) could you summarize these into a short statement of need in the `README`? Such a summary would also be useful to include in the `Description` field of the `DESCRIPTION` file, before the technical summary already there, for people who find the package on CRAN.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1434)._
