### 2019-05-20 #1434 @corybrunson
Regarding @seabbs's miscellaneous comment above, i was able to install **survPen** and reproduce the vignette using R version 3.3.2. I agree that reducing the version dependency is optional, and i wouldn't necessarily recommend it without continuous integration support or unless @fauvernierma can test it using an earlier version on their own machine.
