### 2019-05-20 #1434 @csoneson
@seabbs - thanks for raising the question about the associated methods paper. The software functionality should be possible to review with publicly available material; thus, either the methods paper should be deposited on a preprint server, or enough material to review functionality and performance claims should be included in the documentation of the software.
@fauvernierma - please note that all references that are cited in the paper need to be publicly available (with a DOI or similar identifier).
