+++ .travis.yml
@@ -1,40 +1,15 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+language: r
 
-language: R
-sudo: false
-dist: trusty
-cache: packages
-latex: false
-fortran: true
+sudo: required
 
+env: _R_CHECK_CRAN_INCOMING_=FALSE
 
-r_github_packages: r-lib/testthat
+r_packages:
+  - covr
+  - devtools
 
+r_github_packages:
+  - codecov/example-r
 
-jobs:
-  include:
-  - r: devel
-  - os: osx
-  - r: release
-    after_success:
-    - Rscript -e 'covr::codecov()'
-    before_cache:
-    - Rscript -e 'remotes::install_github("r-lib/pkgdown")'
-    deploy:
-      provider: script
-      script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
-      skip_cleanup: true
-  - r: 3.4
-  - r: 3.3
-  - r: 3.2
-    warnings_are_errors: false
-
-matrix:
-  fast_finish: true
-
-env:
-  global:
-  - _R_CHECK_FORCE_SUGGESTS_=false
-  - MAKEFLAGS="-j 2"
-  - TRAVIS_CXXFLAGS="-Wall -Wextra -pedantic -Werror"
-  - R_REMOTES_NO_ERRORS_FROM_WARNINGS=true
+after_success:
+  - Rscript -e 'library(covr);codecov()'
