### 2019-05-20 2849318
added travis and codecov
