### 2019-05-20 570b8f6
Update .travis.yml
