### 2019-05-21 fauvernierma/survPen#8 @fauvernierma
Ok, thank you both

I implemented some tests using testthat (for now it is pretty basics stuff but it covers the majority of the code) and used travis and codecov to get badges in the README.

Do you think it looks ok ?

Also, to use testhtat with travis I had to add testthat as an import in the description file. Is there any other way ? Typically, when I look at your example code with getTBinR @seabbs , you seem to have done it differently.

Thanks again
