### 2019-05-21 fauvernierma/survPen#8 @corybrunson
@fauvernierma thanks very much for the update.

The new tests confirm that the package is producing the expected kinds of objects (or throws errors when appropriate), and in some key places (excess versus total hazard, concordant penalized likelihoods) confirm that the results are meaningful. They look reasonable to me.

The advantage of producing full output tests like those in `test_survPen.r` is that the complete output can be inspected and compared to expected results. It seems appropriate to keep this file as well. Alternatively, as @seabbs suggested, these calculations could be made into **testthat** scripts that compare freshly-calculated values to "known" values. (For reference, the visual tests could be automated using **vdiffr**, for example, though that'd take some additional getting used to.)

I'm OK either way, since the package now contains both some basic behavioral tests and some manual calculative tests. I'd recommend eventually automating the calculative tests. Though i'd defer to @seabbs's more extensive experience!

Regarding dependency, i've put **testthat** in `Suggests` rather than `Imports`, which should be enough for Travis-CI to attach it when running the tests.
