### 2019-05-24 fauvernierma/survPen#8 @seabbs
These look much better @fauvernierma - good job. 👍

I am also okay with just the base functionality having unit tests with overall functionality having manual tests. It would be great if you could completely automate testing but this isn't a blocker for acceptance for me (as @corybrunson says doing it eventually would be good)

As @corybrunson suggests adding `testthat` (and any other dev package) to suggests is the way to go.

As a final part of testing could you expand your travis grid (and potentially use Appveyor as well) to include more versions of R and other platforms (i.e Mac/Windows). See [here](https://github.com/seabbs/getTBinR/blob/master/.travis.yml) for an example.
