### 2019-05-24 fauvernierma/survPen#7 @seabbs
Just noting that this can be closed if you 1.) want users to submit data or 2.) remove that from the contributing statement.

Have a good weekend :smile:
