### 2019-05-25 fauvernierma/survPen#7 @fauvernierma
The data part from the contributing statement is ok for me

Thanks and good weekend to you too
