### 2019-05-25 cb67af8
added r versions and os
