### 2019-05-27 416a58f
added testthat in Suggests
