+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: Multidimensional Penalized Splines for Survival and Net Survival
 Version: 1.1.1
 Author: Mathieu Fauvernier [aut, cre], Laurent Roche [aut], Laurent Remontet [aut], Zoe Uhry [ctb], Nadine Bossard [ctb]
 Maintainer: Mathieu Fauvernier <mathieu.fauvernier@gmail.com>
-Imports: statmod, stats, testthat
+Imports: statmod, stats
 Description: Fits hazard and excess hazard models with multidimensional penalized splines allowing for time-dependent effects, non-linear effects and interactions between several continuous covariates. 
 Depends: R (>= 3.2.0)
 License: GPL-3 | file LICENSE
@@ -13,7 +13,7 @@ BugReports: https://github.com/fauvernierma/survPen/issues
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
-Suggests: knitr, rmarkdown
+Suggests: knitr, rmarkdown, testthat
 VignetteBuilder: knitr
 NeedsCompilation: no
 Packaged: 2019-05-02 16:09:30 UTC; fauvernierma
