### 2019-05-27 fauvernierma/survPen#8 @fauvernierma
I added testthat in Suggests 416a58f5f8b604e0e1cf624324c5da274f631ebb and updated the travis.yml file 49a7ec2170b31cc54dba1a7c6dfe981a10be49c1

I think it's ok but I am not sure that the packages has been tested on Mac OS. Despite the travis.yml file, it looks like only linux has been taken into account.

Anyway, I guess this test answers another question relative to the R version that survPen actually neeeds (R version 3.2 seems to be a good minimum prerequisite)

Thanks to both of you and have a good day
