### 2019-05-27 fauvernierma/survPen#5 @fauvernierma
I just added the required parts bba311cd201bd5fdd8aa4631671527c059503f6e

If it is ok with you, I will add something similar in the README as well as in the DESCRIPTION file

Thank you
