### 2019-05-28 fauvernierma/survPen#8 @seabbs
Thanks for adding that.

So the default OS on Travis is Linux so you don't need to specify that in the OS section. I'm not an expert in configuring travis but I think if you put `os: osx` that should work. You should see an apple in the travis build list.

It would be great to see the package tested against Windows as well. I think travis now supports this but I generally use Appveyor.

Yes, that is one of the handy benefits! Generally, I follow the tidyverse approach and drop the oldest R version from the testing grid as newer version come out (e.g I recently went from 3.2 -> 3.3 when 3.6 came out).
