### 2019-06-05 #1434 @fauvernierma
Hello @csoneson

The revision has been going quite well so far although I have to admit I did not anticipate the issue with the method paper. Speaking of which, the status of my revised manuscript now reads "With Editor" so it is just a matter of hours until I get the answer.
Hopefully the method paper will be accepted but it is also likely that I will have some other details to fix.

Meanwhile I am finishing my thesis manuscript and the due date is on the 25th of this month and it is going to be pretty complicated for me to work on the JOSS review.

Is it ok if we put it on hold until the 25th ?

Thank you
