### 2019-06-26 #1434 @fauvernierma
Hello everyone,

My method article has finally been accepted for publication in the Journal of the Royal Statistical Society Series C. I will soon be able to provide a DOI.

Concerning the ongoing issues I will respond as soon as possible (probably this Friday or on Monday)

Have a nice day
