### 2019-06-28 fauvernierma/survPen#5 @fauvernierma
Hi, I added the required parts in the README and the DESCRIPTION file
00eed484685070a4fc6faefdacbbe8bd0177d627
799451abe2ae1c58a2fc577cadfaf7aefb7e85c8

Do we consider this issue closed ?
