### 2019-06-28 fauvernierma/survPen#5 @seabbs
These are great additions. However, I think these really resolve #15 rather than this issue. This was more looking at clarifying the first paragraph of the JOSS paper.
