+++ .travis.yml
@@ -14,13 +14,11 @@ r_github_packages:
 jobs:
   include:
   - r: devel
+  - os: osx
+  - r: release
   - r: 3.5
   - r: 3.4
   - r: 3.3
-  - r: 3.2
-  - os:
-    - linux
-    - osx
-  
+
 after_success:
   - Rscript -e 'library(covr);codecov()'
