+++ DESCRIPTION
@@ -8,7 +8,7 @@ Imports: statmod, stats
 Description: Fits hazard and excess hazard models with multidimensional penalized splines allowing for 
 time-dependent effects, non-linear effects and interactions between several continuous covariates. In survival and net survival analysis, in addition to modelling the effect of time (via the baseline hazard), one has often to deal with several continuous covariates and model their functional forms, their time-dependent effects, and their interactions. Model specification becomes therefore a complex problem and penalized regression splines represent an appealing solution to that problem as splines offer the required flexibility while penalization limits overfitting issues. Current implementations of penalized survival models can be slow or unstable and sometimes lack some key features like taking into account expected mortality to provide net survival and excess hazard estimates. In contrast, survPen provides an automated, fast, and stable implementation (thanks to explicit calculation of the derivatives of the likelihood) and offers a unified framework for 
 multidimensional penalized hazard and excess hazard models. survPen may be of interest to those who 1) analyse any kind of time-to-event data: mortality, disease relapse, machinery breakdown, unemployment, etc 2) wish to describe the associated hazard and to understand which predictors impact its dynamics.
-Depends: R (>= 3.2.0)
+Depends: R (>= 3.3.0)
 License: GPL-3 | file LICENSE
 URL: https://github.com/fauvernierma/survPen
 BugReports: https://github.com/fauvernierma/survPen/issues
