### 2019-07-01 c696264
deleting windows (R software not supported)
