### 2019-07-01 fauvernierma/survPen#15 @corybrunson
Thanks! The new text looks good in the README, but it should be shortened to a single paragraph for the `Description` field. The CRAN article on [Writing R Extensions](https://cran.r-project.org/doc/manuals/r-release/R-exts.html#The-DESCRIPTION-file) contains complete details, including this bit:

> The mandatory ‘`Description`’ field should give a _comprehensive_ description of what the package does. One can use several (complete) sentences, but only one paragraph. It should be intelligible to all the intended readership (e.g. for a CRAN package to all CRAN users).
