### 2019-07-02 fauvernierma/survPen#15 @fauvernierma
Hi,

Yes, Travis made me realize this pretty clearly by not wanting to load survPen :)

It is ok now 48ca526b80bc8400db20a24d3321a3fc912fe6a6 I think

Thanks
