### 2019-07-05 fauvernierma/survPen#8 @fauvernierma
Hi, concerning osx it is all ok 4331167ec22ee1c961410c04b510b3038cbc7d97

However, when it comes to windows this is a bit more tricky as R software is not yet available on Travis
I tried to use Appveyor but I am a bit struggling with it. I am not even sure how it works and how to copy and paste someone else's yml file.
Do you happen to have a minimum example .yml file ?

Thanks a lot
