+++ vignettes/survival_analysis_with_survPen.Rmd
@@ -57,9 +57,9 @@ knitr::opts_chunk$set(
 # Introduction
 
 
-The *survPen* package was designed to fit hazard and excess hazard models with multidimensional penalized splines 
+The `survPen` package was designed to fit hazard and excess hazard models with multidimensional penalized splines 
 allowing for time-dependent effects, non-linear effects and interactions between several covariates (Fauvernier et al. 2019). 
-The linear predictor in *survPen* is the logarithm of the (excess) hazard. 
+The linear predictor in `survPen` is the logarithm of the (excess) hazard. 
 
 As the hazard function fully determines the distribution of the time-to-event, this modelling approach is actually 
 well-suited for many time-to-event analyses: the splines provide the flexibility required for modelling the hazard 
@@ -68,29 +68,29 @@ Remontet et al. 2007, Remontet et al. 2018) is linked to the concept of net surv
 and can be useful in specific situations, for example to study the mortality associated with chronic diseases 
 (e.g., cancer survival). 
 
-The framework is very similar to that of the R *mgcv* package developed by Wood for generalized additive models; it 
+The framework is very similar to that of the R `mgcv` package developed by Wood for generalized additive models; it 
 allows including parametric smooth terms based on restricted cubic regression splines as marginal bases, associated 
 with penalties on the second derivative. 
 Multidimensional smoothers are based on tensor product splines, i.e. a term-by-term multiplication of the marginal bases. 
 Smoothing parameters are estimated automatically by optimizing either the Laplace approximate marginal likelihood (LAML) 
 or the likelihood cross-validation criterion (LCV). 
 
-The user must be aware that the *survPen* package is independent of the *mgcv* package and that some functionalities 
-available in *mgcv* in terms of types of splines (such as thin plate regression splines or P-splines) are not 
-available in *survPen* (yet).
+The user must be aware that the `survPen` package is independent of the `mgcv` package and that some functionalities 
+available in `mgcv` in terms of types of splines (such as thin plate regression splines or P-splines) are not 
+available in `survPen` (yet).
 
-In *survPen*, the linear predictor, i.e. the log-hazard function, is fully and explicitly specified by the model’s formula,
+In `survPen`, the linear predictor, i.e. the log-hazard function, is fully and explicitly specified by the model’s formula,
 including the baseline hazard and all time-dependent effects. Thus, time-dependent effects are naturally specified as 
 interactions with functions of time. 
 
-The main functions of the *survPen* package are *survPen*, *smf*, *tensor*, *tint* and *rd*. 
-The *survPen* function fits the model specified in the *formula* argument. The functions *smf*, *tensor*, *tint* are used
-to define penalized splines within this *formula*. Finally, *rd* allows including random effects in the linear predictor.
+The main functions of the `survPen` package are `survPen`, `smf`, `tensor`, `tint` and `rd`. 
+The `survPen` function fits the model specified in the `formula` argument. The functions `smf`, `tensor`, `tint` are used
+to define penalized splines within this `formula`. Finally, `rd` allows including random effects in the linear predictor.
 
-Unpenalized terms can also be incorporated in *survPen* formulae, just as one would specify the terms of a linear predictor 
-in a *glm* formula. The *survPen* package thus allows to easily define and fit various hazard models. As an example, 
-analyses performed using the *coxph* function of the *survival* package, to fit a Cox proportional hazard model, may 
-readily be improved using the *survPen* package by adding a penalized spline to model the baseline hazard: in addition to 
[…]
