### 2019-07-05 fauvernierma/survPen#14 @fauvernierma
Hi,

I understand your point. As for me, it felt important that all functions are exported so the users can get a bit of information. However, as you noted most functions do not have examples because I felt it was pointless to give any (as these functions will never get called directly by the users).

Honestly I am not sure what is best: not to export all functions or add examples for all functions ?

Could it possibly stay as is (like a compromise between the two)

Do you or @seabbs have a preference ?

Thanks
