### 2019-07-05 fauvernierma/survPen#13 @corybrunson
Yes, i found the readability much improved. Thank you!
