### 2019-07-05 fauvernierma/survPen#14 @corybrunson
I just have a preference for exported functions to include examples. ; ) I think i see where you're coming from, but to my mind a user gains little from having access to a function if the behavior of that function is cannot be validated (as against an example). Put another way: Presumably you can imagine a user using these functions directly for some reason. Whatever that reason, could you include an example of it?

Perhaps this is more true for some of the helper functions than others, so perhaps only those few would benefit from examples. I can take another close look in a few days if you'd like more specific feedback—it's not fresh in my mind and i don't have the time this weekend! And i also welcome thoughts from @seabbs.
