### 2019-07-07 #1434 @fauvernierma
Hi,

Things are going in the right direction. A couple of issues were closed this week and some of the main issues are going to be: concerning Automated test (#8), the only question left is how to test the package on windows ; concerning Performance claims (#10), I will send the proofs for my method article this week along with its associated code. The code will be freely available on the JRSSC website so this should settle the Performance claims issue.
