+++ R/survPenV1_2.r
@@ -1,0 +1,4522 @@
+#--------------------------------------------------------------------------------------------------------------------------
+# Penalized (excess) hazard model for time-to-event data
+#--------------------------------------------------------------------------------------------------------------------------
+
+
+#----------------------------------------------------------------------------------------------------------------
+# datCancer : simulated cancer dataset
+#----------------------------------------------------------------------------------------------------------------
+
+#' Patients diagnosed with cervical cancer
+#'
+#' A simulated dataset containing the follow-up times of 2000 patients diagnosed with cervical cancer between 
+#' 1990 and 2010. End of follow-up is June 30th 2013. The variables are as follows:
+#' \itemize{
+#'   \item begin. beginning of follow-up. For illustration purposes about left truncation only (0--1)
+#'   \item fu. follow-up time in years (0--5)
+#'   \item age. age at diagnosis in years, from 21.39 to 99.33
+#'   \item yod. decimal year of diagnosis, from 1990.023 to 2010.999
+#'   \item dead. censoring indicator (1 for dead, 0 for censored)
+#'   \item rate. expected mortality rate (from overall mortality of the general population) (0--0.38)
+#' }
+#' @docType data
+#' @keywords datasets
+#' @name datCancer
+#' @usage data(datCancer)
+#' @format A data frame with 2000 rows and 6 variables
+NULL
+
+
+#----------------------------------------------------------------------------------------------------------------
+# END of code : datCancer
+#----------------------------------------------------------------------------------------------------------------
+
+
+#----------------------------------------------------------------------------------------------------------------
+# survPenObject : description of the object returned by function survPen
+#----------------------------------------------------------------------------------------------------------------
+
+#' Fitted survPen object
+#'
+#' A fitted survPen object returned by function \code{\link{survPen}} and of class "survPen". 
+#' Method functions predict and summary are available for this class.
+#'
+#' @return A \code{survPen} object has the following elements:
+#' \item{call}{original \code{survPen} call}
+#' \item{formula}{formula object specifying the model}
+#' \item{t0.name}{name of the vector of origin times}
+#' \item{t1.name}{name of the vector of follow-up times}
+#' \item{event.name}{name of the vector of right-censoring indicators}
+#' \item{expected.name}{name of the vector of expected hazard}
+#' \item{haz}{fitted hazard}
+#' \item{coefficients}{estimated regression parameters. Unpenalized parameters are first, followed by the penalized ones}
+#' \item{type}{"net" for net survival estimation with penalized excess hazard model or "overall" for overall survival with penalized hazard model}
+#' \item{df.para}{degrees of freedom associated with fully parametric terms (unpenalized)}
+#' \item{df.smooth}{degrees of freedom associated with penalized terms}
+#' \item{p}{number of regression parameters}
+#' \item{edf}{effective degrees of freedom}
+#' \item{edf2}{effective degrees of freedom corrected for smoothing parameter uncertainty}
+#' \item{aic}{Akaike information criterion with number of parameters replaced by edf when there are penalized terms. Corresponds to 2*edf - 2*ll.unpen}
[…]
