+++ README.Rmd
@@ -74,7 +74,12 @@ citation("survPen")
 ```
 
 You may also consider citing the method article:
-Fauvernier, M., Roche, L., Uhry, Z., Tron, L., Bossard, N., Remontet, L. and the CENSUR Working Survival Group (2019). Multidimensional penalized hazard model with continuous covariates: applications for studying trends and social inequalities in cancer survival, in revision in the Journal of the Royal Statistical Society, series C.
+Fauvernier, M., Roche, L., Uhry, Z., Tron, L., Bossard, N., Remontet, L. and the CENSUR Working Survival Group (in press). Multi-dimensional penalized hazard model with continuous covariates: applications for studying trends and social inequalities in cancer survival, 
+Journal of the Royal Statistical Society, series C. doi: 10.1111/rssc.12368
+
+
+
+
 
 
 
