### 2019-07-09 fauvernierma/survPen#14 @seabbs
I completely agree with @corybrunson.

 Personally, I think everything should have an example unless there is a clear reason for it not to. This helps users understand the workings of each function in the package rather than treating it as a complete black box.

A benefit of having examples for each function is that they can relatively easily be converted into meaningful tests. The tests as currently stands are mostly checking the class of objects and this would go a long way to improving them.
