### 2019-07-09 fauvernierma/survPen#8 @seabbs
Yes try using this one: https://github.com/seabbs/getTBinR/blob/master/appveyor.yml

I'd be very keen on the unit tests being expanded to be more meaningful + less automated. Looking at them it looks like quite a few of them are just checking the class etc. which isn't really testing anything (I may be wrong here :)).
