### 2019-07-09 e9aa9a6
added appveyor badge
