+++ R/survPenV1_2.r
@@ -135,6 +135,16 @@ NULL
 #' @return Matrix of dimensions n*(p1*p2) representing the row tensor product of the matrices X1 and X2
 #' @export
 #'
+#' @examples
+#'
+#' library(survPen)
+#'
+#' set.seed(15)
+#'
+#' X1 <- matrix(rnorm(10*3),nrow=10,ncol=3)
+#' X2 <- matrix(rnorm(10*2),nrow=10,ncol=2)
+#' tensor.in(X1,X2)
+#' 
 tensor.in <- function(X1,X2){
 
 	# each column of X1 is multiplied by all the columns of X2
@@ -161,6 +171,17 @@ tensor.in <- function(X1,X2){
 #' \item{T}{Matrix of dimensions n*(p1*p2*...*pm) representing the row tensor product of the matrices in X}
 #' @export
 #'
+#' @examples
+#'
+#' library(survPen)
+#'
+#' set.seed(15)
+#'
+#' X1 <- matrix(rnorm(10*3),nrow=10,ncol=3)
+#' X2 <- matrix(rnorm(10*2),nrow=10,ncol=2)
+#' X3 <- matrix(rnorm(10*2),nrow=10,ncol=2)
+#' tensor.prod.X(list(X1,X2,X3))
+#' 
 tensor.prod.X <- function (X) 
 {
     m <- length(X) # number of matrices
@@ -207,6 +228,19 @@ tensor.prod.X <- function (X)
 #' \item{TS}{List of the penalty matrices associated with the tensor product smooth}
 #' @export
 #'
+#' @examples
+#'
+#' library(survPen)
+#'
+#' set.seed(15)
+#'
+#' S1 <- matrix(rnorm(3*3),nrow=3,ncol=3)
+#' S2 <- matrix(rnorm(2*2),nrow=2,ncol=2)
+#' 
+#' S1 <- 0.5*(S1 + t(S1) ) ; S2 <- 0.5*(S2 + t(S2) )
+#'
+#' tensor.prod.S(list(S1,S2))
+#' 
 tensor.prod.S <- function (S) 
 {
     m <- length(S)
@@ -409,6 +443,15 @@ crs <- function(x, knots=NULL,df=10, intercept=TRUE) {
 #' \item{P.mat}{penalty matrix}
 #' @export
 #'
+#' @examples
+#'
+#' library(survPen)
+#'
+#' knots <- c(0,0.25,0.5,0.75,1)
+#' diff.knots <- diff(knots)
+#'
+#' crs.FP(knots,diff.knots)
+#' 
 crs.FP <- function(knots,h){
   # constraints of second derivatives continuity and nullity beyond the boundaries
 
@@ -614,6 +657,16 @@ rd <- function(...){
 #' \item{name}{simplified name of the call to function smooth.spec}
 #' @export
 #'
+#' @examples
+#' 
+#' library(survPen)
+#'
+#' # standard spline of time with 10 unspecified knots
+#' smooth.spec(time)
+#'
+#' # tensor of time and age with 5*5 specified knots
+#' smooth.s <- smooth.spec(time,age,knots=list(time=seq(0,5,length=5),age=seq(20,80,length=5)), option="tensor")
+#'
 smooth.spec <- function(..., knots=NULL,df=NULL,by=NULL,option=NULL,same.rho=FALSE){
 
   if (is.null(option)) {
@@ -752,6 +805,16 @@ smooth.spec <- function(..., knots=NULL,df=NULL,by=NULL,option=NULL,same.rho=FAL
 #' \item{lambda.name}{name of the smoothing parameters}
 #' @export
 #'
+#' @examples
+#' 
+#' library(survPen)
+#'
[…]
