### 2019-07-09 fauvernierma/survPen#8 @fauvernierma
ok thanks

It seems to be ok e9aa9a6bf8d3db6b2a6d907d2bf2a757d8e9f802

I have also added examples for all functions but three of them (inv.repam, cor.var and print.summary, these three are therefore not exported any longer)

I will work on the new examples to provide better tests

Thanks
