### 2019-07-09 fauvernierma/survPen#14 @fauvernierma
I added examples 086e829433695a0cbd28c9d9f72099a610c2d327 for all functions but three of them (inv.repam, cor.var and print.summary, these three are therefore not exported any longer)
