### 2019-07-09 fauvernierma/survPen#12 @fauvernierma
Hi,

Binder seems to be working 3c7517677b6f593a42bebec781c0e9ad6684288e

However I find it a bit slow, is it normal in your experience ?
This can be related to the connection I have at work though :)
