### 2019-07-09 fauvernierma/survPen#12 @seabbs
Looks good to me. Thanks for adding this - a little bit of a personal passion for me.

Lagging or slow compute? Compute will be slow as they assign you a very small amount of a vCPU. Shouldn't be too laggy though.
