### 2019-07-09 fauvernierma/survPen#12 @seabbs
Just run through the vignette and yes it is pretty slow. Not something for serious work but good for discovery.
