### 2019-07-09 2340ff4
Update test_tensor.R
