### 2019-07-09 6184eef
Create test_Newton.R
