### 2019-07-09 2883a76
Update test_tensor.R
