### 2019-07-09 6c04e75
Update test_Newton.R
