+++ .travis.yml
@@ -18,7 +18,6 @@ jobs:
   - r: release
   - r: 3.5
   - r: 3.4
-  - r: 3.3
 
 after_success:
   - Rscript -e 'library(covr);codecov()'
