### 2019-07-10 e251d65
Update test_smf.R
