### 2019-07-10 bbf0820
Update test_random_effects.R
