### 2019-07-10 fauvernierma/survPen#9 @fauvernierma
Hi,

Most of what was required concerning this issue has been dealt with and there seems to be two main points left : the website and the smaller vignettes.
To me, these two points feel a bit minor relative to what the documentation has to offer. Concerning the only vignette I understand it can look a bit intimidating but it is exhaustive and its format (html) makes it easy to navigate into the different sections. That is the reason why I find redundant and tiresome to split the current vignette into smaller ones. And to be honest I feel the same about the website (to me the website is the github page).
