### 2019-07-10 fauvernierma/survPen#9 @seabbs
The following are things that I expect a good modern package to have:

1. A quick start guide in the README
2. A comparison of the package compared to others linked from the documentation or in the documentation.
3. Package website

Happy for you to have a single vignette. Personally, I think this may scare off some users but this is a personal choice on your part.

A `pkgdown` site is great as it gives a much better place to explore function documentation. As they are so standard across packages it also makes it easier for users to explore the functionality of your package.

What do you think @corybrunson? Happy to skip any/all.
