### 2019-07-10 fauvernierma/survPen#14 @seabbs
Great. It would be good to see full comments as far as possible.

I am not sure but I think you can skip loading `survPen` in every example.
