### 2019-07-10 fauvernierma/survPen#14 @fauvernierma
ok thanks

I would rather keep the loading part as I often forget to load it myself :)
I think it is convenient to be able to copy and paste it for every example but I do not know if it is good practice
