### 2019-07-10 fauvernierma/survPen#8 @fauvernierma
I added more tests mostly concerning errors b871687e7a4af7c6b27d023b327980430778c243, the internal Newton algorithms 6c04e755dbe531154d1b33ebade959c65d9b0d0a and some expected prediction values aabfb8b4b91149c720ea6fc95c0eefc6b769c5431

Hopefully this is closer to what you expected. But don't hesitate to tell me if it is not :)

Thank you
