### 2019-07-10 fauvernierma/survPen#9 @corybrunson
I agree with @seabbs about the utility of including, e.g., a quick start in the README as well as the sort of thorough vignette @fauvernierma has already written. But i think we're also on the same page about the criteria for recommendation and acceptance to JOSS. The documentation is there, so i consider that element of the review process complete. The developers (or users!) can augment it with intros or shorter vignettes in future.

However, if the package makes claims to improved performance over alternatives (see #10), then these must be substantiated. My understanding is that this will have been done in the paper, once published. (It would then of course need to be available in some form for free and hyperlinked from the README or vignette.) Alternatively, the demonstration could be part of the package, with expository text and output along with system specs, which would most naturally be done as a vignette.
