### 2019-07-10 fauvernierma/survPen#9 @fauvernierma
Thank you both for these pieces of advice

@corybrunson, concerning the performance claims, we totally agree. I have just returned the corrected proofs along with the code associated with the paper. As soon as the Journal gives me the final link to access the code I will add it into github

Thanks again
