### 2019-07-10 fauvernierma/survPen#8 @corybrunson
The testthat tests now test the structure of the output and the behavior of `print()` and `summary()` methods as well as classes, but also—most importantly for me—some of the numerical results (namely, the predictions). It would be good to also test the coefficient estimates, as was done by the original test files, but i think i'm morally satisfied with the current tests. Thank you @fauvernierma for adding these!

@seabbs it may be that statistical model-fitting software really should include specific tests that aren't here yet, and i'd like to defer to you on that—this also being an issue you originally raised, of course. : )
