### 2019-07-10 70c43f4
Update survPenV1_2.r
