### 2019-07-10 fauvernierma/survPen#14 @fauvernierma
I just added some comments for each example 70c43f44cc8329c3541f95d8578cc74c1f705d5f
