### 2019-07-11 fauvernierma/survPen#8 @seabbs
Thanks for adding those @fauvernierma - much improved.

My only remaining concern is that so many of the tests are just checking inherited class. I may be wrong but is that not just checking that any given function runs without an error?

The issue with this is that it inflates the test coverage and won't protect you (or other developers) should something change down the line. Personally, I would like to see these tests dropped/replaced with some more meaningful tests. What do you think @corybrunson

Otherwise, I am happy with the current state of the tests - good job. Adding more tests would be great (really encourage you to add as many as possible either now or as development continues :smile:) but the current set is much more thorough than many packages have.

Depending on what @corybrunson thinks about class tests I am happy to consider this done. 👍
