### 2019-07-11 fauvernierma/survPen#8 @fauvernierma
Hi @seabbs, you are right, the inherited class test is a way to check if an error occurred.

They are clearly not perfect that's why I added the prediction test file which is much more meaningful. And of course we agree on the fact that this is still an ongoing process :)
