+++ R/survPenV1_2.r
@@ -79,7 +79,7 @@ NULL
 #' \item{Hess.beta}{hessian of the penalized log-likelihood with respect to the regression parameters}
 #' \item{Hess.beta.modif}{if TRUE, the hessian of the penalized log-likelihood has been perturbed at convergence}
 #' \item{ll.unpen}{log-likelihood at convergence}
-#' \item{ll}{penalized log-likelihood at convergence}
+#' \item{ll.pen}{penalized log-likelihood at convergence}
 #' \item{deriv.rho.beta}{transpose of the Jacobian of beta with respect to the log smoothing parameters}
 #' \item{deriv.rho.inv.Hess.beta}{list containing the derivatives of the inverse of \code{Hess} with respect to the log smoothing parameters}
 #' \item{deriv.rho.Hess.unpen.beta}{list containing the derivatives of \code{Hess.unpen} with respect to the log smoothing parameters}
@@ -3179,7 +3179,7 @@ survPen.fit <- function(build,data,formula,max.it.beta=200,beta.ini=NULL,detail.
   beta <- Algo.optim$beta
   names(beta) <- colnames(X)
   ll.unpen <- Algo.optim$ll.unpen
-  ll <- Algo.optim$ll
+  ll.pen <- Algo.optim$ll.pen
   haz.GL <- Algo.optim$haz.GL
   iter.beta <- Algo.optim$iter.beta
   #-------------------------------------------------------------------
@@ -3301,7 +3301,7 @@ survPen.fit <- function(build,data,formula,max.it.beta=200,beta.ini=NULL,detail.
 	log.det.Hess.beta <- as.numeric(2*determinant(R,logarithm=TRUE)$modulus)
 
 	# this is actually the negative LAML so that we can minimize it
-	LAML <- -(ll+0.5*log.abs.S-0.5*log.det.Hess.beta+0.5*M.p*log(2*pi))
+	LAML <- -(ll.pen+0.5*log.abs.S-0.5*log.det.Hess.beta+0.5*M.p*log(2*pi))
 
 	if(method=="LAML"){
 
@@ -3687,7 +3687,7 @@ survPen.fit <- function(build,data,formula,max.it.beta=200,beta.ini=NULL,detail.
 	haz=ft1,coefficients=beta,type=type,df.para=df.para,df.smooth=df.smooth,p=p,edf=edf,edf2=edf2,aic=2*sum(edf)-2*ll.unpen,aic2=aic2,iter.beta=iter.beta,X=X,S=S,S.scale=S.scale,
 	S.list=S.list,S.smf=S.smf,S.tensor=S.tensor,S.tint=S.tint,S.rd=S.rd,smooth.name.smf=smooth.name.smf,smooth.name.tensor=smooth.name.tensor,smooth.name.tint=smooth.name.tint,smooth.name.rd=smooth.name.rd,
 	S.pen=build$S.pen,grad.unpen.beta=grad.unpen.beta,grad.beta=grad.beta,Hess.unpen.beta=Hess.unpen.beta,Hess.beta=Hess.beta,
-	Hess.beta.modif=Hess.beta.modif,ll.unpen=ll.unpen,ll=ll,deriv.rho.beta=deriv.rho.beta,deriv.rho.inv.Hess.beta=deriv.rho.inv.Hess.beta,lambda=lambda,
+	Hess.beta.modif=Hess.beta.modif,ll.unpen=ll.unpen,ll.pen=ll.pen,deriv.rho.beta=deriv.rho.beta,deriv.rho.inv.Hess.beta=deriv.rho.inv.Hess.beta,lambda=lambda,
[…]
