### 2019-07-12 ebedf44
Update test_smf.R
