### 2019-07-12 5002213
Update survival_analysis_with_survPen.R
