### 2019-07-12 fauvernierma/survPen#10 @fauvernierma
Hello to you both,

So unless I missed something, this issue about Performance Claims should be the only one left ?

As I said in another relative issue, the details to reproduce the code from my method article will be available on the Journal website. It will normally be on this page : https://rss.onlinelibrary.wiley.com/hub/journal/14679876/series-c-datasets

and then in "Volume 68 (2019), part 5" (not yet available)
