### 2019-07-12 fauvernierma/survPen#10 @seabbs
Thanks @fauvernierma

Waiting on what @corybrunson thinks about the documentation and test issues. As I said in those issues I would like a quick start in the README, a package website and more honest tests (i.e not testing for errors/not only testing for errors).

Will the paper be open access when it goes up. To me, it looked like my institution was providing access? I think that JOSS guidelines will need the performance claims to be substantiated by open access information.
