### 2019-07-12 fauvernierma/survPen#10 @fauvernierma
The paper will not be open access but I think that the code will be (I can't access all articles on the website but I had no problem downloading code so far). Therefore with enough information on github in the README file this should be easy for people to reproduce the comparison with the other packages (as well as the simulation study from the article)
