### 2019-07-18 #1434 @corybrunson
Hi, i apologize for being absent for several days—i had an unreliable connection over the weekend but i should have mentioned that here and affirmed that i'd be back soon. I'm returning to the outstanding issues now.
