### 2019-07-18 fauvernierma/survPen#10 @corybrunson
I believe that open-access code demonstrating the performance claims will satisfy this criterion! I don't want to check it off preemptively, though, so shall we revisit this issue when the article is published?
