### 2019-07-18 fauvernierma/survPen#14 @corybrunson
@fauvernierma could you update the package documentation? The new examples don't appear there yet.
