### 2019-07-19 fauvernierma/survPen#10 @fauvernierma
Hi,

Indeed, I guess it is simpler if we wait for the article to be published
