+++ man/summary.survPen.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{summary.survPen}
 \alias{summary.survPen}
 \title{Summary for a \code{survPen} fit}
@@ -31,4 +31,20 @@ List of objects:
 }
 \description{
 Takes a fitted \code{survPen} object and produces various useful summaries from it.
+}
+\examples{
+
+library(survPen)
+
+data(datCancer) # simulated dataset with 2000 individuals diagnosed with cervical cancer
+
+# model : unidimensional penalized spline for time since diagnosis with 5 knots
+f1 <- ~smf(fu,df=5)
+
+# fitting hazard model
+mod1 <- survPen(f1,data=datCancer,t1=fu,event=dead,expected=NULL,method="LAML")
+
+# summary
+summary(mod1)
+
 }
