+++ man/survPen.fit.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{survPen.fit}
 \alias{survPen.fit}
 \title{(Excess) hazard model with multidimensional penalized splines for given smoothing parameters}
@@ -29,4 +29,27 @@ Object of class "survPen" (see \code{\link{survPenObject}} for details)
 }
 \description{
 Fits an (excess) hazard model. If penalized splines are present, the smoothing parameters are specified.
+}
+\examples{
+
+library(survPen)
+
+# standard spline of time with 4 knots
+
+data <- data.frame(time=seq(0,5,length=100),event=1,t0=0)
+
+form <- ~ smf(time,knots=c(0,1,3,5))
+
+t1 <- eval(substitute(time), data)
+t0 <- eval(substitute(t0), data)
+event <- eval(substitute(event), data)
+	
+# Setting up the model before fitting
+model.c <- model.cons(form,lambda=0,data.spec=data,t1=t1,t1.name="time",
+t0=rep(0,100),t0.name="t0",event=event,event.name="event",
+expected=NULL,expected.name=NULL,type="overall",n.legendre=20,cl="survPen(form,data,t1=time,event=event)")
+ 
+# fitting
+mod <- survPen.fit(model.c,data,form)
+
 }
