+++ man/NR.beta.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{NR.beta}
 \alias{NR.beta}
 \title{Inner Newton-Raphson algorithm for regression parameters estimation}
@@ -22,7 +22,7 @@ NR.beta(build, beta.ini, detail.beta, max.it.beta = 200,
 List of objects:
 \item{beta}{estimated regression parameters}
 \item{ll.unpen}{log-likelihood at convergence}
-\item{ll}{penalized log-likelihood at convergence}
+\item{ll.pen}{penalized log-likelihood at convergence}
 \item{haz.GL}{list of all the matrix-vector multiplications X.GL[[i]]\%*\%beta for Gauss Legendre integration. Useful to avoid repeating operations in \code{\link{survPen.fit}}}
 \item{iter.beta}{number of iterations needed to converge}
 }
@@ -32,7 +32,30 @@ convergence : first the hessian is perturbed whenever it is not positive definit
 log-likelihood is not maximized, the step is halved until it is.
 }
 \details{
-If we note \code{ll} and \code{beta} respectively the current penalized log-likelihood and estimated parameters and
-\code{llold} and \code{betaold} the previous ones, the algorithm goes on while
-(abs(ll-llold)>tol.beta) or any(abs((beta-betaold)/betaold)>tol.beta)
+If we note \code{ll.pen} and \code{beta} respectively the current penalized log-likelihood and estimated parameters and
+\code{ll.pen.old} and \code{betaold} the previous ones, the algorithm goes on while
+(abs(ll.pen-ll.pen.old)>tol.beta) or any(abs((beta-betaold)/betaold)>tol.beta)
+}
+\examples{
+
+library(survPen)
+
+# standard spline of time with 4 knots
+
+data <- data.frame(time=seq(0,5,length=100),event=1,t0=0)
+
+form <- ~ smf(time,knots=c(0,1,3,5))
+
+t1 <- eval(substitute(time), data)
+t0 <- eval(substitute(t0), data)
+event <- eval(substitute(event), data)
+	
+# Setting up the model before fitting
+model.c <- model.cons(form,lambda=0,data.spec=data,t1=t1,t1.name="time",
+t0=rep(0,100),t0.name="t0",event=event,event.name="event",
+expected=NULL,expected.name=NULL,type="overall",n.legendre=20,cl="survPen(form,data,t1=time,event=event)")
+ 
+# Estimating the regression parameters at given smoothing parameter (here lambda=0)
+Newton1 <- NR.beta(model.c,beta.ini=rep(0,4),detail.beta=TRUE)
+
 }
