+++ man/tensor.in.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{tensor.in}
 \alias{tensor.in}
 \title{tensor model matrix for two marginal bases}
@@ -16,4 +16,16 @@ Matrix of dimensions n*(p1*p2) representing the row tensor product of the matric
 }
 \description{
 Function called recursively inside \code{\link{tensor.prod.X}}.
+}
+\examples{
+
+library(survPen)
+
+# row-wise tensor product between two design matrices
+set.seed(15)
+
+X1 <- matrix(rnorm(10*3),nrow=10,ncol=3)
+X2 <- matrix(rnorm(10*2),nrow=10,ncol=2)
+tensor.in(X1,X2)
+
 }
