+++ man/design.matrix.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{design.matrix}
 \alias{design.matrix}
 \title{Design matrix for the model needed in Gauss-Legendre quadrature}
@@ -32,4 +32,31 @@ design matrix for the model
 \description{
 Builds the design matrix for the whole model when the sum-to-zero constraints are specified. The function is called inside \code{\link{model.cons}}
 for Gauss-Legendre quadrature.
+}
+\examples{
+
+library(survPen)
+
+# standard spline of time with 4 knots
+
+data <- data.frame(time=seq(0,5,length=100),event=1,t0=0)
+
+form <- ~ smf(time,knots=c(0,1,3,5))
+
+t1 <- eval(substitute(time), data)
+t0 <- eval(substitute(t0), data)
+event <- eval(substitute(event), data)
+	
+# Setting up the model
+model.c <- model.cons(form,lambda=0,data.spec=data,t1=t1,t1.name="time",
+t0=rep(0,100),t0.name="t0",event=event,event.name="event",
+expected=NULL,expected.name=NULL,type="overall",n.legendre=20,cl="survPen(form,data,t1=time,event=event)")
+ 
+# Retrieving the sum-to-zero constraint matrices and the list of knots
+Z.smf <- model.c$Z.smf ; list.smf <- model.c$list.smf
+
+# Calculating the design matrix
+design.M <- design.matrix(form,data.spec=data,Z.smf=Z.smf,list.smf=list.smf,Z.tensor=NULL,Z.tint=NULL,
+list.tensor=NULL,list.tint=NULL,list.rd=NULL)
+
 }
