+++ man/print.summary.survPen.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{print.summary.survPen}
 \alias{print.summary.survPen}
 \title{print summary for a \code{survPen} fit}
