+++ man/NR.rho.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{NR.rho}
 \alias{NR.rho}
 \title{Outer Newton-Raphson algorithm for smoothing parameters estimation via LCV or LAML optimization}
@@ -50,4 +50,28 @@ is not minimized, the step is halved until it is.
 If we note \code{val} the current LCV or LAML value,
 \code{val.old} the previous one and \code{grad} the gradient vector of LCV or LAML with respect to the log smoothing parameters, the algorithm goes on
 \code{while(abs(val-val.old)>tol.rho|any(abs(grad)>tol.rho))}
+}
+\examples{
+
+library(survPen)
+
+# standard spline of time with 4 knots
+
+data <- data.frame(time=seq(0,5,length=100),event=1,t0=0)
+
+form <- ~ smf(time,knots=c(0,1,3,5))
+
+t1 <- eval(substitute(time), data)
+t0 <- eval(substitute(t0), data)
+event <- eval(substitute(event), data)
+	
+# Setting up the model before fitting
+model.c <- model.cons(form,lambda=0,data.spec=data,t1=t1,t1.name="time",
+t0=rep(0,100),t0.name="t0",event=event,event.name="event",
+expected=NULL,expected.name=NULL,type="overall",n.legendre=20,cl="survPen(form,data,t1=time,event=event)")
+ 
+# Estimating the smoothing parameter and the regression parameters
+# we need to apply a reparameterization to model.c before fitting
+Newton2 <- NR.rho(repam(model.c)$build,rho.ini=-1,data,form,nb.smooth=1,detail.rho=TRUE)
+
 }
