+++ man/cor.var.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/survPenV1_10.r
+% Please edit documentation in R/survPenV1_2.r
 \name{cor.var}
 \alias{cor.var}
 \title{Implementation of the corrected variance Vc}

[…]
