### 2019-07-19 fauvernierma/survPen#14 @corybrunson
Yes, they're showing up now. Thanks!

I ran into at least one problem, though, in the `model.cons()` example that also appears as a test:

``` r
library(survPen)

data <- data.frame(time=seq(0,5,length=100),event=1,t0=0)

form <- ~ smf(time,knots=c(0,1,3,5))

t1 <- eval(substitute(time), data)
t0 <- eval(substitute(t0), data)
event <- eval(substitute(event), data)

model.c <- model.cons(form,lambda=0,data.spec=data,t1=t1,t1.name="time",
                      t0=rep(0,100),t0.name="t0",event=event,event.name="event",
                      expected=NULL,expected.name=NULL,type="overall",n.legendre=20,cl="survPen(form,data,t1=time,event=event)")
#> Error in model.cons(form, lambda = 0, data.spec = data, t1 = t1, t1.name = "time", : dims [product 400] do not match the length of object [0]
```

Does this run successfully for others?

(To caveat, i've been running the code on R v3.3.3, so i've changed `DESCRIPTION` locally to allow installation. I doubt that's the issue, but i mention it just in case.)
