### 2019-07-19 fauvernierma/survPen#14 @fauvernierma
There is no problem for me with R version >= 4.0.0 so I guess version could be an issue here
