### 2019-07-19 fauvernierma/survPen#14 @fauvernierma
ok

The "expected" argument should not be NULL here but "rep(0,100)". Otherwise, at the end of the function model.cons, the following expression (line 1979 of the survPen code)

eventXexpected <- eventX*expected

returns an error. R version < 4.0.0 does not accept the multiplication between a matrix and a NULL whereas R version >=4.0.0 accepts this
