### 2019-07-19 a58768b
Update test_Newton.R
