### 2019-07-19 fauvernierma/survPen#14 @fauvernierma
To support my explanation see
https://cran.r-project.org/bin/windows/base/old/3.4.0/NEWS.R-3.4.0.html

"In integer arithmetic, NULL is now treated as integer(0) whereas it was previously treated as double(0)."
