+++ README.Rmd
@@ -1,3 +1,5 @@
+
+
 ---
 output: github_document
 ---
@@ -12,9 +14,7 @@ knitr::opts_chunk$set(
 
 # survPen
 ## hazard and excess hazard modelling with multidimensional penalized splines
-[![badge](https://img.shields.io/badge/Launch-survPen-blue.svg)](https://mybinder.org/v2/gh/fauvernierma/survPen/master?urlpath=rstudio)
 [![Travis-CI Build Status](https://travis-ci.org/fauvernierma/survPen.svg?branch=master)](https://travis-ci.org/fauvernierma/survPen)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/fauvernierma/survPen?branch=master&svg=true)](https://ci.appveyor.com/project/fauvernierma/survpen)
 [![Coverage Status](https://img.shields.io/codecov/c/github/fauvernierma/survPen/master.svg)](https://codecov.io/github/fauvernierma/survPen?branch=master)
 [![DOI](https://zenodo.org/badge/181266005.svg)](https://zenodo.org/badge/latestdoi/181266005)
 
@@ -50,16 +50,97 @@ or directly from the CRAN repository:
 install.packages("survPen")
 ```
 
-and then use it by typing
-
-```{r loading, eval = FALSE}
-library(survPen)
-```
 
 ## The functionalities of the `survPen` package are extensively detailed in the following vignette
 Available from
 [GitHub](https://htmlpreview.github.io/?https://github.com/fauvernierma/survPen/blob/master/inst/doc/survival_analysis_with_survPen.html)
 or from [CRAN](https://cran.r-project.org/web/packages/survPen/vignettes/survival_analysis_with_survPen.html)
+
+
+## Quick start
+
+In the following section, we will use a simulated dataset that contains artificial data from 2,000 women 
+diagnosed with cervical cancer between 1990 and 2010. End of follow-up is June 30th 2013.
+
+Let's fit a penalized hazard model and a penalized excess hazard model. Without covariates for now.
+```{r}
+library(survPen)
+
+f1 <- ~ smf(fu) # penalized natural cubic spline of time with 10 knots placed at the quantiles
+
+mod.total <- survPen(f1,data=datCancer,t1=fu,event=dead,method="LAML")
+
+mod.excess <- survPen(f1,data=datCancer,t1=fu,event=dead,expected=rate,method="LAML")
+```
+
+Now let's look at the predicted dynamics of the hazard (and excess hazard) as well as the predicted 
+survival curve (and net survival curve)
+
+```{r plot-haz, fig.height = 8}
+lwd1 = 2
+
+# compare the predictions of the models
+new.time <- seq(0,5,length=100)
+pred.total <- predict(mod.total,data.frame(fu=new.time))
+pred.excess <- predict(mod.excess,data.frame(fu=new.time))
+
+# hazard vs excess hazard
+par(mfrow=c(1,2))
+plot(new.time,pred.total$haz,type="l",ylim=c(0,0.2),main="hazard vs excess hazard",
+xlab="time since diagnosis (years)",ylab="hazard",lwd=lwd1)
+lines(new.time,pred.excess$haz,col="red",lwd=lwd1,lty=2)
+legend("topright",legend=c("total","excess"),col=c("black","red"),lty=c(1,2), lwd=rep(lwd1,2))
+
+plot(new.time,pred.total$surv,type="l",ylim=c(0,1),main="survival vs net survival",
+xlab="time",ylab="survival",lwd=lwd1)
+lines(new.time,pred.excess$surv,col="red",lwd=lwd1,lty=2)
[…]
