+++ README.md
@@ -129,7 +129,7 @@ summary(mod.tensor)
 #> converged= TRUE
 ```
 
-Now let's predict the predicted hazard surface
+Now let's predict the predicted excess hazard surface
 
 
 
