### 2019-07-19 fauvernierma/survPen#14 @corybrunson
Aha, OK, then not only does the example work, but the default value `expected = NULL` is appropriate for R `v>=3.4.0`. Since this is a key point, i would suggest leaving `expected = NULL` in the test but changing it to `expected = rep(0, 100)` in the example. Or, to make the package more backward-compatible, the function could be amended to handle `expected = NULL` more explicitly. But that's not a requirement! The issue with `model.cons()` is basically resolved.

I also haven't come across any other problems with the examples or tests, so, since i opened this issue (and since @seabbs has given the go-ahead), i'll close it. Thanks for addressing these questions @fauvernierma !
