### 2019-07-19 fauvernierma/survPen#9 @corybrunson
The Quick Start is immensely helpful! Thanks for that @fauvernierma. My only suggestion is to add a concluding sentence after the excess hazard surface, to make the takeaway explicit and to help separate the quick start from the next section (contributing). (This is not a requirement for recommendation, just a stylistic suggestion.)
