### 2019-07-25 fauvernierma/survPen#10 @corybrunson
Congrats @fauvernierma ! I'll have a look at the paper and try out the comparison code ASAP.

By the way, if there are any references to this "forthcoming" paper in the survPen documentation, have you updated them?
