### 2019-07-25 fauvernierma/survPen#10 @fauvernierma
Thanks !

There is a reference at the end of the README. I will update it to inform people that they can try out the comparison code

By the way, starting from tomorrow I will be on vacation until August 19 so I will be able to see your messages but I will not always be able to update the github page
