### 2019-07-30 fauvernierma/survPen#10 @corybrunson
I'm not versed in the comparison packages but the demonstration script `0000_Compar_packages.R` clearly suggests that survPen scales (in terms of combinatorial tensor effects) better than the comparator implementations. It would be nice to have the comparisons looped, so that they could be more easily and compactly expanded (e.g. to include different data sets or additional combinations of parameter settings) but that's more a suggestion for future work than an expectation for acceptance to JOSS. I think i'm satisfied.

This is @seabbs 's issue, so i'll leave it to them to make the determination. : )
