### 2019-07-30 #1434 @corybrunson
@csoneson i'm sorry for the delay on my part in reviewing the last component i was waiting for, the comparison to other implementations. @fauvernierma has now been able to make this public, so my checklist is complete, but i'd like to wait until @seabbs has commented on that issue to give my recommendation.
