### 2019-07-30 fauvernierma/survPen#10 @seabbs
Congrats on the paper @fauvernierma! I agree that the comparison scripts do the job but also agree with @corybrunson suggestions for improvements.

I'd like to see these comparisons more clearly linked in the documentation. Ideally, as I have commented previously, these would be in the documentation but I am happy to leave that for a future update.  So I think the only work left to do for JOSS acceptance is some clear links in at least the README and also potentially the vignette and paper.
