### 2019-07-30 fauvernierma/survPen#8 @seabbs
Do you think the current testing suite is sufficient @corybrunson? I noted my concerns above but am happy to close this and leave them to later updates if you think that makes sense?
