### 2019-07-30 fauvernierma/survPen#9 @seabbs
Quickstart is a really nice addition thanks @fauvernierma. The final thing is a link to the performance claims in the documentation.

Website is optional (but encouraged).
