### 2019-07-30 #1434 @seabbs
I am just waiting on a few small changes to the documentation and need to check @corybrunson thoughts on the testing suite. Once both those are done I will be happy to recommend this for acceptance.

Thanks for checking in @csoneson

Great job @fauvernierma
