### 2019-07-30 fauvernierma/survPen#9 @seabbs
Sorry just noticed that both the binder and appveyor badges have been dropped from the README - could you clarify why @fauvernierma?
