### 2019-07-30 fauvernierma/survPen#8 @corybrunson
@seabbs thanks for checking. : ) I think my take on this issue is the same as on the performance demo, in that i have some suggestions for improvements but that they rise to a level slightly higher than the acceptance requirements. The real demonstration of a package's utility is that people use it, and i'll bet that use cases will more precisely indicate the sorts of tests that are needed.
