+++ README.md
@@ -1,10 +1,11 @@
 
 # survPen
 ## hazard and excess hazard modelling with multidimensional penalized splines
+[![badge](https://img.shields.io/badge/Launch-survPen-blue.svg)](https://mybinder.org/v2/gh/fauvernierma/survPen/master?urlpath=rstudio)
 [![Travis-CI Build Status](https://travis-ci.org/fauvernierma/survPen.svg?branch=master)](https://travis-ci.org/fauvernierma/survPen)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/fauvernierma/survPen?branch=master&svg=true)](https://ci.appveyor.com/project/fauvernierma/survpen)
 [![Coverage Status](https://img.shields.io/codecov/c/github/fauvernierma/survPen/master.svg)](https://codecov.io/github/fauvernierma/survPen?branch=master)
 [![DOI](https://zenodo.org/badge/181266005.svg)](https://zenodo.org/badge/latestdoi/181266005)
-
 
 In survival and net survival analysis, in addition to modelling the effect of time (via the baseline hazard), 
 one has often to deal with several continuous covariates and model their functional forms, their time-dependent 
@@ -199,9 +200,14 @@ citation("survPen")
 ```
 
 You may also consider citing the method article:
-Fauvernier, M., Roche, L., Uhry, Z., Tron, L., Bossard, N., Remontet, L. and the CENSUR Working Survival Group (in press). Multi-dimensional penalized hazard model with continuous covariates: applications for studying trends and social inequalities in cancer survival, 
+Fauvernier, M., Roche, L., Uhry, Z., Tron, L., Bossard, N., Remontet, L. and the CENSUR Working Survival Group (2019). Multi-dimensional penalized hazard model with continuous covariates: applications for studying trends and social inequalities in cancer survival, 
 Journal of the Royal Statistical Society, series C. doi: 10.1111/rssc.12368
 
+The article and its supplementary are available here
+https://rss.onlinelibrary.wiley.com/doi/full/10.1111/rssc.12368
+
+The supplementary provides a comparison between `survPen` and existing approaches. The code associated with this comparison and with the simulation study from the article is available here
+https://github.com/fauvernierma/code_Fauvernier_JRSSC_2019
 
 
 
