### 2019-07-31 fauvernierma/survPen#10 @fauvernierma
Hello to you both,

I just added the necessary information in the README bd70d4178efc46e64e4eb0f056909adeb294d9c1

Please let me know if anything is missing

Thank you
