### 2019-07-31 #1434 @seabbs
@fauvernierma has completed the final changes so my checklist is complete and all review issues are closed. I am now very happy to recommend this for acceptance @csoneson

Congrats 🎆  @fauvernierma  - this is a really nice package. Looking forward to making use of it in the future.

Hopefully, the review process has been useful!
