+++ paper/paper.bib
@@ -33,7 +33,8 @@
 @article{Fauvernier2019,
 	author = {Fauvernier, M. and Roche, L. and Uhry, Z. and Tron, L. and Bossard, N. and Remontet, L. and the CENSUR working survival group},
 	title = {Multidimensional penalized hazard model with continuous covariates: applications for studying trends and social inequalities in cancer survival},
-	journal = {in revision in the Journal of the Royal Statistical Society Series C Applied Statistics},
+	journal = {Journal of the Royal Statistical Society Series C Applied Statistics},
+	DOI = {10.1111/rssc.12368},
 	year = {2019},
 	type = {Journal Article}
 }
