### 2019-07-31 #1434 @csoneson
@seabbs, @corybrunson - great, thanks a lot!
@fauvernierma - I will have another look through to see if there are any formalities that need to be fixed before I accept the paper.
