### 2019-07-31 1a3ab01
Fix minor formatting issues
