+++ paper/paper.bib
@@ -11,14 +11,13 @@
 	0140-6736 (Linking)},
 	DOI = {10.1016/S0140-6736(17)33326-3},
 	url = {https://www.ncbi.nlm.nih.gov/pubmed/29395269},
-	year = {2018},
-	type = {Journal Article}
+	year = {2018}
 }
 
 @article{Esteve1990,
 	author = {Estève, J. and Benhamou, E. and Croasdale, M. and Raymond, L.},
 	title = {Relative survival and the estimation of net survival: elements for further discussion},
-	journal = {Stat Med},
+	journal = {Statistics in Medicine},
 	volume = {9},
 	number = {5},
 	pages = {529-38},
@@ -26,8 +25,7 @@
 	0277-6715 (Linking)},
 	DOI = {10.1002/sim.4780090506},
 	url = {https://www.ncbi.nlm.nih.gov/pubmed/2349404},
-	year = {1990},
-	type = {Journal Article}
+	year = {1990}
 }
 
 @article{Fauvernier2019,
@@ -35,21 +33,19 @@
 	title = {Multidimensional penalized hazard model with continuous covariates: applications for studying trends and social inequalities in cancer survival},
 	journal = {Journal of the Royal Statistical Society Series C Applied Statistics},
 	DOI = {10.1111/rssc.12368},
-	year = {2019},
-	type = {Journal Article}
+	year = {2019}
 }
 
 @article{OSullivan1988,
 	author = {O’Sullivan, Finbarr},
 	title = {Fast computation of fully automated log-density and log-hazard estimators},
-	journal = {SIAM Journal on scientific and statistical computing},
+	journal = {SIAM Journal on Scientific and Statistical Computing},
 	volume = {9},
 	number = {2},
 	pages = {363-379},
 	ISSN = {0196-5204},
 	DOI = {10.1137/0909024},
-	year = {1988},
-	type = {Journal Article}
+	year = {1988}
 }
 
 @article{Perme2012,
@@ -61,8 +57,7 @@
 	pages = {113-120},
 	ISSN = {0006-341X},
 	DOI = {10.1111/j.1541-0420.2011.01640.x},
-	year = {2012},
-	type = {Journal Article}
+	year = {2012}
 }
 
 
@@ -79,14 +74,13 @@
 @article{Remontet2007,
 	author = {Remontet, Laurent and Bossard, Nadine and Belot, Aurélien and Estève, Jacques and the French Network of Cancer Registries},
 	title = {An overall strategy based on regression models to estimate relative survival and model the effects of prognostic factors in cancer survival studies},
-	journal = {Statistics in medicine},
+	journal = {Statistics in Medicine},
 	volume = {26},
 	number = {10},
 	pages = {2214-2228},
 	ISSN = {0277-6715},
 	DOI = {10.1002/sim.2656},
-	year = {2007},
-	type = {Journal Article}
+	year = {2007}
 }
 
 
@@ -95,8 +89,7 @@
 	title = {Semiparametric Regression},
 	publisher = {Cambridge University Press},
 	year = {2003},
-	DOI = {10.1017/CBO9780511755453},
-	type = {Book}
+	DOI = {10.1017/CBO9780511755453}
 }
 @article{Uhry2017,
 	author = {Uhry, Z. and Bossard, N. and Remontet, L. and Iwaz, J. and Roche, L. and the Grell Eurocare-5 Working Group and the Censur Working Survival Group},
@@ -108,16 +101,14 @@
 	0959-8278 (Linking)},
 	DOI = {10.1097/CEJ.0000000000000301},
 	url = {https://www.ncbi.nlm.nih.gov/pubmed/28005600},
-	year = {2017},
[…]
