### 2019-07-31 #1434 @csoneson
@fauvernierma - I sent a PR with a couple of small formatting fixes; please check and merge if you agree, then run `@whedon generate pdf` to generate a new version of the paper.

In addition to that, the next thing to do is to make a tagged release of your software, and create a Zenodo archive. Please make sure that the Zenodo archive has exactly the same title and authors as the paper. Once that's done, please report back here with the DOI of the Zenodo archive.
