### 2019-08-02 #1434 @fauvernierma
Hello, @csoneson

I would like to make "the CENSUR working survival group" appear as a group author after the name "Laurent Roche", is it possible ?

I thought I did it but obviously it did not work

Thank you
