### 2019-08-02 #1434 @csoneson
Hi @fauvernierma - sorry, just to clarify, when you say "is it possible", and "it did not work" - do you mean from a technical point of view, or is the question whether it is ok to have a group author?
