### 2019-08-03 #1434 @csoneson
I'm sorry, I'm probably missing the point, but could you not just add "the CENSUR working survival group" as another author in the markdown yaml header?
