### 2019-08-03 #1434 @csoneson
@fauvernierma - could you please clarify why you would like to add a group author at this stage, and what are the contributions to the submission? According to the JOSS authorship guidelines (https://joss.readthedocs.io/en/latest/submitting.html#authorship), purely financial or organizational contributions, for example, are not sufficient for inclusion among the authors.
