### 2019-08-05 #1434 @csoneson
Thank you for the clarification! To me this sounds like an "organizational contribution"; however I would like to check with the current AEiC on duty for their opinion as well @openjournals/joss-eics

>I would like to add the contribution of a working group that all the other authors and I are a part of. This CENSUR project involves our team and other teams from other European countries. It is cited as an author for the method article as it offers guidelines and subjects for future work and the method article and survPen are related to one of said subject.
>Do you think it is possible to add the CENSUR project as an author ? I totally understand if it is not
