### 2019-08-06 #1434 @arfon
> Thank you for the clarification! To me this sounds like an "organizational contribution"; however I would like to check with the current AEiC on duty for their opinion as well @openjournals/joss-eics

I think that's OK. I'm not exactly how our editorial toolchain will handle this but perhaps you can just add an extra author called `CENSUR project team` with an affiliation of `CENSUR project` or something like that?
