### 2019-08-08 #1434 @fauvernierma
I want to check with them because I am not used to adding an affiliation for this kind of group author
