### 2019-08-20 #1434 @fauvernierma
Hi,

It looks almost good to us, I just need to make sure that the CENSUR group won't be considered as the last author.

Typically, if you look at this article in pubmed
https://www.ncbi.nlm.nih.gov/pubmed?term=(Roche%5BAuthor%20-%20Last%5D)%20AND%20%22Statistical%20methods%20in%20medical%20research%22%5BJournal%5D

the CENSUR group appears at the end as a "CORPORATE AUTHOR" but Laurent Roche is considered as the last author.

Can you confirm that for the JOSS article, Laurent Roche will be considered as the last author ?

Thank you very much
