### 2019-08-20 3f7a482
Update paper.md
