### 2019-08-20 #1434 @arfon
As currently written, the CENSUR group is the last author. This PR should fix that: https://github.com/fauvernierma/survPen/pull/17
