### 2019-08-20 #1434 @fauvernierma
Hello again,

I apologize if my request seems weird but this is not really an ordering. The CENSUR group should be treated differently than the "proper" authors.

For example, for the article I cited before, the information concering the article is indexed as follows

https://www.ncbi.nlm.nih.gov/pubmed?term=(Flexible%5BTitle%5D%20AND%20structured%5BTitle%5D%20AND%20survival%5BTitle%5D%20AND%20model%5BTitle%5D%20AND%20simultaneous%5BTitle%5D%20AND%20estimation%5BTitle%5D%20AND%20non-linear%5BTitle%5D%20AND%20non-proportional%5BTitle%5D%20AND%20effects%5BTitle%5D%20AND%20complex%5BTitle%5D%20AND%20interactions%5BTitle%5D%20AND%20continuous%5BTitle%5D%20AND%20variables%5BTitle%5D%20AND%20Performance%5BTitle%5D%20AND%20multidimensional%5BTitle%5D%20AND%20penalized%5BTitle%5D%20AND%20spline%5BTitle%5D%20AND%20approach%5BTitle%5D%20AND%20net%5BTitle%5D%20AND%20survival%5BTitle%5D%20AND%20trend%5BTitle%5D%20AND%20analysis%5BTitle%5D)&report=medline&format=text

As we can see, CENSUR is not an author (AU) but a collective name (CN). As a matter of fact, it is not even associated with a affiliation

Sorry again for this weird request. If this does not meet the JOSS criteria, maybe it is best to forget about CENSUR.

Thank you
