### 2019-08-20 #1434 @arfon
Ah OK, thanks for clarifying. I'm afraid this isn't something we can support with our paper metadata.

It sounds like it might be best to remove CENSUR from the author list.
