### 2019-08-20 #1434 @fauvernierma
ok it is done (https://github.com/fauvernierma/survPen/commit/0ac67f698ca7a9eb540107a55685e8b557a998e3)

I will create the zenodo archive now

Thank you for your time
