### 2019-08-21 #1434 @csoneson
@fauvernierma - one final thing. I went through the paper again and found a typo; I fixed it and sent a PR.
