### 2019-08-21 #1434 @fauvernierma
Ok I just merged it

Thank you @csoneson
