### 2019-08-21 #1434 @csoneson
@openjournals/joss-eics - we are ready to accept this paper!
