### 2019-08-21 #1434 @kyleniemeyer
The final acceptance is up to the the JOSS editors-in-chief team—I will do some final checks first.
