+++ paper/paper.bib
@@ -29,9 +29,9 @@
 }
 
 @article{Fauvernier2019,
-	author = {Fauvernier, M. and Roche, L. and Uhry, Z. and Tron, L. and Bossard, N. and Remontet, L. and the CENSUR working survival group},
+	author = {Fauvernier, M. and Roche, L. and Uhry, Z. and Tron, L. and Bossard, N. and Remontet, L. and {the Challenges in the Estimation of Net Survival Working Survival Group}},
 	title = {Multidimensional penalized hazard model with continuous covariates: applications for studying trends and social inequalities in cancer survival},
-	journal = {Journal of the Royal Statistical Society Series C Applied Statistics},
+	journal = {Journal of the Royal Statistical Society: Series C (Applied Statistics)},
 	DOI = {10.1111/rssc.12368},
 	year = {2019}
 }
