### 2019-08-21 b51df6a
Update paper.md
Changes a few in-text references
