### 2019-08-21 #1434 @kyleniemeyer
@fauvernierma  nearly there! We just need a few changes in the paper:

- One of our paper requirements is:
> Begin your paper with a summary of the high-level functionality of your software for a non-specialist reader. Avoid jargon in this section.

Could you add this at the beginning of your paper, introducing also what survival and net survival analysis are?
- I made two small PRs for the paper and bibliography, if you could merge those: https://github.com/fauvernierma/survPen/pull/19 and https://github.com/fauvernierma/survPen/pull/20
