### 2019-08-23 342f66c
Update paper.md
