### 2019-08-23 #1434 @fauvernierma
Hello @kyleniemeyer,

I performed some changes in the article. Let me know if that is what you had in mind

Thank you
