### 2019-08-23 #1434 @kyleniemeyer
Congrats @fauvernierma on your article's publication in JOSS! Many thanks to @csoneson for editing, and @corybrunson and @seabbs for reviewing, this submission.
