### 2019-08-26 #1434 @fauvernierma
Hi,

Thanks a lot everyone. I am looking forward to submitting another package to JOSS !

Have a great day !!!!
